% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_create_reservation.R
\name{api_create_reservation}
\alias{api_create_reservation}
\title{Create reservation}
\usage{
api_create_reservation(scope, environment, user.id, user.pass, affiliation)
}
\arguments{
\item{scope}{(character) Scope of identifier to be reserved (e.g. edi, knb-lter-ntl).}

\item{environment}{(character) Data repository environment in which to reserve the
identifier. Can be: 'development', 'staging', 'production'.}

\item{user.id}{(character) Identification of user reserving the identifier.}

\item{user.pass}{(character) Password corresponding with the user.id argument supplied
above.}

\item{affiliation}{(character) Affiliation corresponding with the user.id argument supplied
above. Can be: 'LTER' or 'EDI'.}
}
\value{
(character) Package identifier.
}
\description{
Create Reservation operation, creates a new reservation in PASTA for the 
    specified user on the next reservable identifier for the specified 
    scope. The integer value of the reserved identifier (as assigned by 
    PASTA) is returned in the web service response body. User authentication 
    is required.
}
