% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_list_data_descendants.R
\name{api_list_data_descendants}
\alias{api_list_data_descendants}
\title{List data descendants}
\usage{
api_pkg_data_descendants(package.id, environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('XMLInternalDocument' 'XMLAbstractDocument') XML containing the the 
    data package ID, data package title, and data package URL.
}
\description{
List Data Descendants operation, specifying the scope, identifier, 
    and revision values to match in the URI. Data descendants are data 
    packages that are known to be derived, in whole or in part, from 
    the specified source data package.
}
