% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_list_data_package_scopes.R
\name{api_list_data_package_scopes}
\alias{api_list_data_package_scopes}
\title{List data package scopes}
\usage{
api_list_data_package_scopes(environment = 'production')
}
\arguments{
\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) All scope values extant in the data package registry.
}
\description{
List data package scopes
}
