% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_list_deleted_data_packages.R
\name{api_list_deleted_data_packages}
\alias{api_list_deleted_data_packages}
\title{List deleted data packages}
\usage{
api_list_data_package_scopes(environment = 'production')
}
\arguments{
\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Deleted data packages
}
\description{
List Deleted Data Packages operation, returning all document identifiers 
    (excluding revision values) that have been deleted from the data package 
    registry.
}
