% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_list_principal_owner_citations.R
\name{api_list_principal_owner_citations}
\alias{api_list_principal_owner_citations}
\title{List principal owner citations}
\usage{
api_list_principal_owner_citations(dn, environment = 'production')
}
\arguments{
\item{dn}{(character) Distinguished name (e.g. 
'uid=csmith,o=LTER,dc=ecoinformatics,dc=org')}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('XMLInternalDocument' 'XMLAbstractDocument') journalCitations metadata
}
\description{
List Principal Owner Citations operation.
}
