% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_list_recent_uploads.R
\name{api_list_recent_uploads}
\alias{api_list_recent_uploads}
\title{List recent uploads}
\usage{
api_list_recent_uploads(type, limit = 5, environment = 'production')
}
\arguments{
\item{type}{(character) Upload type. Can be: "insert" or "update".}

\item{limit}{(integer) Maximum limit.}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('XMLInternalDocument' 'XMLAbstractDocument') dataPackageUploads 
    metadata.
}
\description{
List Recent Uploads operation, optionally specifying the upload type 
    (“insert” or “update”) and a maximum limit as query parameters in the 
    URL.
}
