% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read_data_entity_sizes.R
\name{api_read_data_entity_sizes}
\alias{api_read_data_entity_sizes}
\title{Read data entity sizes}
\usage{
api_read_data_entity_sizes(
      package.id, 
      environment = 'production'
    )
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1', 'knb-lter-ntl.303.20).}

\item{environment}{(character) PASTA+ environment in which to perform the operation. Can 
be: 'development', 'staging', 'production'.}
}
\value{
(data frame) A data frame containing entity identifiers and entity 
    sizes (in bytes). Each line contains an entity identifier and its 
    corresponding size value. Only data entities that the user is authorized 
    to read are included.
}
\description{
Read Data Entity Sizes operation, specifying the scope, identifier, and 
    revision of the data package whose data entity sizes are to be 
    retrieved.
}
