% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read_data_package_report.R
\name{api_read_data_package_report}
\alias{api_read_data_package_report}
\title{Read data package report}
\usage{
api_read_data_package_report(package.id, environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('xml_document' 'xml_node') Data package report.
}
\description{
Read Data Package Report operation, specifying the scope, identifier, 
    and revision of the data package quality report document to be read 
    in the URI.
}
