% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read_metadata_format.R
\name{api_read_metadata_format}
\alias{api_read_metadata_format}
\title{Read metadata format}
\usage{
api_read_metadata_format(package.id, environment = 
    'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) The metadata format type, e.g. 
    “eml://ecoinformatics.org/eml-2.1.1”
}
\description{
Read Metadata Format operation, specifying the scope, identifier, and 
    revision of the metadata to be read in the URI.
}
