% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_evaluate_data_package.R
\name{auth_key}
\alias{auth_key}
\title{Make authentication key}
\usage{
auth_key(user.id, affiliation)
}
\arguments{
\item{user.id}{(character) Identification of user performing the evaluation.}

\item{affiliation}{(character) Affiliation corresponding with the user.id argument supplied
above. Can be: 'LTER' or 'EDI'.}
}
\description{
Create user authentication key for PASTA+ operations.
}
