% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{authenticate}
\alias{authenticate}
\title{Authenticate with EDI credentials}
\usage{
authenticate(config = NULL)
}
\arguments{
\item{config}{(character) Path to config.txt containing a user name and password (see details below). If NULL, the console will prompt for this information.}
}
\value{
(file) A temporary (~18 hour) authentication token written to edi_token.txt within the session \code{tempdir()}.
}
\description{
Authenticate with EDI credentials
}
\details{
\code{config}: Supplying credentials in a file named config.txt, and containing the lines "userId = NAME" and "userPass = PASS", facilitates authentication within automated/unassisted processes.

Security: Input user name and password are only held for the duration of this function call and are disposed of by \code{on.exit()}. The temporary encrypted token is written to file at the path defined by \code{tempdir()} and is removed at close of the users R session.
}
\examples{
\dontrun{
# Authenticate at console
authenticate()

# Authenticate with config.txt
authenticate("/Users/me/Documents/config.txt")
}

}
