% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{bake_cookie}
\alias{bake_cookie}
\title{Create authentication cookie for EDI repository Gatekeeper}
\usage{
bake_cookie()
}
\value{
(request) The request object returned by \code{httr::set_cookies()} with the EDI repository authentication token baked in. Yum!
}
\description{
Create authentication cookie for EDI repository Gatekeeper
}
