% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{base_url}
\alias{base_url}
\title{Construct base URL of the EDI repository web services}
\usage{
base_url(env)
}
\arguments{
\item{env}{(character) Data repository environment to perform the evaluation in. Can be: 'development', 'staging', 'production'.}
}
\description{
Construct base URL of the EDI repository web services
}
