% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_evaluation_status.R
\name{check_evaluation_status}
\alias{check_evaluation_status}
\title{Check data package evaluation status}
\usage{
check_evaluation_status(transaction, tier = "production")
}
\arguments{
\item{transaction}{(character) Transaction identifier}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(list) A list containing:
\itemize{
  \item "evaluated" (logical) TRUE if evaluation has completed and the report is available
  \item "response" (response) Response from status check on \code{transaction}
}
}
\description{
Check data package evaluation status
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{
path <- "/Users/me/Documents/edi.468.1.xml"
transaction <- evaluate_data_package(path)
check_evaluation_status(transaction)
}

}
