% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_eml.R
\name{compare_node_as_string}
\alias{compare_node_as_string}
\title{Collapse EML node and compare as string}
\usage{
compare_node_as_string(newest, previous, xpath)
}
\arguments{
\item{newest}{(xml_document, xml_node) EML of the newest version of a data package, where inputs are returned from \code{api_read_metadata()}.}

\item{previous}{(xml_document, xml_node) EML of the previous version of a data package, where inputs are returned from \code{api_read_metadata()}.}

\item{xpath}{(character) xpath to node of interest}
}
\value{

}
\description{
Collapse EML node and compare as string
}
