% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_tables.R
\name{compare_tables}
\alias{compare_tables}
\title{Did a set of tables change between versions?}
\usage{
compare_tables(newest, previous)
}
\arguments{
\item{newest}{(list) Tables from the newest version of a data package, where inputs are returned from \code{read_tables()}.}

\item{previous}{(list) Tables from the previous version of a data package, where inputs are returned from \code{read_tables()}.}
}
\value{
(character) A vector of check results
}
\description{
If table attributes changed, then downstream processes relying on an expected structure may fail. This checks for common structural changes in the tables (not EML, see \code{compare_eml()}).
}
\examples{

compare_tables(
  newest = read_tables(api_read_metadata("knb-lter-hfr.118.32")),
  previous = read_tables(api_read_metadata("knb-lter-hfr.118.31")))

}
