% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{config_test_eml}
\alias{config_test_eml}
\title{Set environment variables for testing data package evaluation and upload}
\usage{
config_test_eml(userId, url)
}
\arguments{
\item{userId}{(character) EDI repository userId}

\item{url}{(character) URL from which the EDI repository can download the test data.txt object. This URL cannot contain any redirects.}
}
\value{
Environmental variables \code{EDI_USERID = userId} and \code{EDI_TEST_URL = url}
}
\description{
Testing data package evaluation and upload requires a web accessible data object, EML metadata describing the data object, and an EDI repository user account. Use of this function presupposes the data object has been stashed
}
\details{
The results of this function are used to create a test EML file for create, update, and delete tests.
}
