% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tables.R
\name{convert_missing_value}
\alias{convert_missing_value}
\title{Convert missing value codes to NA}
\usage{
convert_missing_value(v, code, type)
}
\arguments{
\item{v}{Vector of values}

\item{code}{(character) Missing value code}

\item{type}{(character) Type (class) \code{v} should be. Supported types are: "character", "numeric", "datetime"}
}
\value{
Vector of values with \code{code} replaced by NA in the class of \code{type}
}
\description{
Convert missing value codes to NA
}
