% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_audit_record.R
\name{create_audit_record}
\alias{create_audit_record}
\title{Create audit record}
\usage{
create_audit_record(tier = "production")
}
\arguments{
\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
New logged entry in the Audit Manager’s logging database
}
\description{
Create audit record
}
\examples{

}
