% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_staged.R
\name{data_package_citation_exists}
\alias{data_package_citation_exists}
\title{Check if the staged package has linked journal citations}
\usage{
data_package_citation_exists(staged_id, all_versions = T)
}
\arguments{
\item{staged_id}{(character) The full package identifier (scope, accession, revision) of an EDI Data Package in the PASTA staging environment}

\item{all_versions}{(boolean) If true, checks all revisions of the staged package. If false, only checks the specified revision}
}
\value{
\code{TRUE} if the staged package has a linked journal citation, \code{FALSE} if not.
}
\description{
Check if the staged package has linked journal citations
}
\examples{

# Check if staged data package edi.101.1 has a linked journal citation

data_pacakge_citation_exists('edi.101.1')

}
