% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_data_package.R
\name{delete_data_package}
\alias{delete_data_package}
\title{Delete data package}
\usage{
delete_data_package(scope, identifier, env = "production")
}
\arguments{
\item{scope}{(character) Scope of data package}

\item{identifier}{(numeric) Identifier of data package}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(logical) TRUE if deleted
}
\description{
Delete data package
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{

login()

# Create data package
transaction <- create_data_package(
 eml = "./data/edi.595.1.xml", 
 env = "staging")
transaction
#> [1] "create_163966765080210573"

# Check creation status
status <- check_status_create(
 transaction = transaction, 
 packageId = "edi.595.1", 
 env = "staging")
status
#> [1] TRUE

# Delete data package
delete_data_package(
  scope = "edi", 
  identifier = "595",
  env = "staging")
#> [1] TRUE

logout()
}

}
