% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{enter_polling_loop}
\alias{enter_polling_loop}
\title{Enter polling loop}
\usage{
enter_polling_loop(transaction)
}
\arguments{
\item{transaction}{(character) Transaction identifier}
}
\value{

}
\description{
Enter polling loop
}
\examples{

}
