% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datetime_format.R
\name{get_datetime_format}
\alias{get_datetime_format}
\title{Get date time format string}
\usage{
get_datetime_format(x)
}
\arguments{
\item{x}{(character) A vector of dates and times created with
`EDIutils::datetime_to_iso8601`.}
}
\value{
(character) A datetime format string if only one is found, otherwise
    a vector of datetime format strings are returned.
}
\description{
Get format of ISO 8601 formatted character string.
}
