% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_distinguished_name}
\alias{get_distinguished_name}
\title{Get distinguished name}
\usage{
get_distinguished_name(uid, o)
}
\arguments{
\item{uid}{(character) PASTA userId}

\item{o}{(character) Organizational unit in which \code{uid} belongs. Can be "EDI" or "LTER".}
}
\value{
(character) Distinguished name
}
\description{
Get distinguished name
}
\examples{
get_distinguished_name("csmith", "EDI")

}
