% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eml_attribute.R
\name{get_eml_attribute}
\alias{get_eml_attribute}
\title{Get attribute metadata}
\usage{
get_eml_attribute(attr.name, package.id)
}
\arguments{
\item{attr.name}{(character) Name of an attribute found in an EML metadata record}

\item{package.id}{(character) ID of data package found in the Environmental Data
Initiative repository (e.g. knb-lter-cap.627.3), containing the
attribute defined by "attr.name".}
}
\value{
A named vector of metadata elements for the specified "attr.name". NA is
    returned when the metadata element doesn't exist.
}
\description{
Retrieve metadata for an attribute
}
