% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_physical.R
\name{get_physical}
\alias{get_physical}
\title{Get data entity physical metadata}
\usage{
get_physical(eml, entityId)
}
\arguments{
\item{eml}{(xml_document) EML metadata}

\item{entityId}{(character) Data entity identifier}
}
\value{
(xml_node) The physical metadata of \code{entityId} as described in \code{eml}
}
\description{
Get data entity physical metadata
}
\examples{
entityId <- "f6e4efd0b04aea3860724824ca05c5dd"
eml <- read_metadata("knb-lter-cap.691.2")
phys <- get_physical(eml, entityId)
phys

}
