% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provenance_metadata.R
\name{get_provenance_metadata}
\alias{get_provenance_metadata}
\title{Get provenance metadata}
\usage{
get_provenance_metadata(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Provenance metadata of \code{packageId}, representing a <methodStep> element that can be inserted into the <methods> section of a dependent data package
}
\description{
Generates the provenance metadata of a source data package
}
\examples{
get_provenance_metadata("knb-lter-pal.309.1")

}
