% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_active_reservations.R
\name{list_active_reservations}
\alias{list_active_reservations}
\title{List active reservations}
\usage{
list_active_reservations(as = "data.frame", env = "production")
}
\arguments{
\item{as}{(character) Format of the returned object. Can be: "data.frame" 
or "xml".}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(data.frame or xml_document) The set of data package identifiers 
that users have actively reserved. Note that data package identifiers that 
have been successfully uploaded are no longer considered active reservations 
and thus are not included in this list.
}
\description{
List active reservations
}
\examples{
\dontrun{

# List reservations
reservations <- list_active_reservations()
}
}
\seealso{
Other Identifier Reservations:
\code{\link[=create_reservation]{create_reservation()}},
\code{\link[=delete_reservation]{delete_reservation()}},
\code{\link[=list_reservation_identifiers]{list_reservation_identifiers()}}
}
\concept{Identifier Reservations}
