% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_scopes.R
\name{list_data_package_scope}
\alias{list_data_package_scope}
\title{List data package scopes}
\usage{
list_data_package_scopes(environment = 'production')
}
\arguments{
\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) All scope values extant in the data package registry.
}
\description{
List data package scopes
}
