% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_scopes.R
\name{list_data_package_scopes}
\alias{list_data_package_scopes}
\title{List data package scopes}
\usage{
list_data_package_scopes(env = "production")
}
\arguments{
\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(numeric) Scopes within a specified \code{env}
}
\description{
List data package scopes
}
\examples{
\dontrun{

# List scopes
scopes <- list_data_package_scopes()
scopes
#>  [1] "ecotrends"           "edi"                 "knb-lter-and"       
#>  [4] "knb-lter-arc"        "knb-lter-bes"        "knb-lter-ble"       
#>  [7] "knb-lter-bnz"        "knb-lter-cap"        "knb-lter-cce"       
#> [10] "knb-lter-cdr"        "knb-lter-cwt"        "knb-lter-fce"       
#> [13] "knb-lter-gce"        "knb-lter-hbr"        "knb-lter-hfr"       
#> [16] "knb-lter-jrn"        "knb-lter-kbs"        "knb-lter-knz"       
#> [19] "knb-lter-luq"        "knb-lter-mcm"        "knb-lter-mcr"       
#> [22] "knb-lter-nes"        "knb-lter-nin"        "knb-lter-ntl"       
#> [25] "knb-lter-nwk"        "knb-lter-nwt"        "knb-lter-pal"       
#> [28] "knb-lter-pie"        "knb-lter-sbc"        "knb-lter-sev"       
#> [31] "knb-lter-sgs"        "knb-lter-vcr"        "lter-landsat"       
#> [34] "lter-landsat-ledaps" "msb-cap"             "msb-paleon"         
#> [37] "msb-tempbiodev"   
}
}
\seealso{
Other Listing:
\code{\link[=list_data_descendants]{list_data_descendants()}},
\code{\link[=list_data_entities]{list_data_entities()}},
\code{\link[=list_data_package_identifiers]{list_data_package_identifiers()}},
\code{\link[=list_data_package_revisions]{list_data_package_revisions()}},
\code{\link[=list_data_sources]{list_data_sources()}},
\code{\link[=list_deleted_data_packages]{list_deleted_data_packages()}},
\code{\link[=list_recent_changes]{list_recent_changes()}},
\code{\link[=list_recent_uploads]{list_recent_uploads()}},
\code{\link[=list_service_methods]{list_service_methods()}},
\code{\link[=list_user_data_packages]{list_user_data_packages()}}
}
\concept{Listing}
