% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to the EDI repository}
\usage{
login(
  userId = NULL,
  userPass = NULL,
  key = NULL,
  config = NULL,
  env = "production"
)
}
\arguments{
\item{userId}{(character) User identifier of an EDI data repository account.}

\item{userPass}{(character) Password of \code{userId}}

\item{key}{(character) EDI-API access key.}

\item{config}{(character) Path to config.txt, which contains \code{userId}
and \code{userPass} or \code{key} (see details below)}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(NULL) No return value. The token or API key is written to its respective environment variable.
}
\description{
Login to the EDI repository
}
\details{
If \code{userId}, \code{userPass}, \code{key}, and \code{config} are NULL, the console
will prompt for credentials.

\code{config}: Supplying credentials in a file named config.txt facilitates
authentication within automated/unassisted processes. Contents of this file
should be new line separated and have the form "<argument> = <value>" (e.g.
userId = myname or key = my_api_key).
}
\note{
Legacy credentials login only works when authenticating with EDI credentials. 
For modern token-free authentication, generate an API key from the EDI Portal 
and supply it as the \code{key} parameter or in your \code{config} file.

Be careful not to accidentally share your credentials.
Some tips to avoid this:
\itemize{
  \item Don't write code that explicitly lists your credentials.
  \item Don't save your workspace when exiting an R session.
  \item Do store your credentials as environmental variables and reference
  these.
  \item Do use \code{config} but if using version control ensure the
  config.txt file is listed in your .gitignore.
}
If you may have shared your credentials, please reset your password at
\url{https://dashboard.edirepository.org/dashboard/auth/reset_password_init}.
}
\examples{
\dontrun{

# Interactively at the console
login()
#> EDI-API key (leave blank to use username/password): "my_api_key"

# Programmatically with an API key
login(key = "my_api_key")

# Programmatically with legacy function arguments
login(userId = "my_name", userPass = "my_secret", env = "production")

# Programmatically with a file containing credentials
login(config = paste0(tempdir(), "/config.txt"))
}

}
\seealso{
Other Authentication:
\code{\link[=logout]{logout()}}
}
\concept{Authentication}
