% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_packageId}
\alias{parse_packageId}
\title{Parse package ID into scope, identifier, and revision}
\usage{
parse_packageId(package.id)
}
\arguments{
\item{package.id}{(character) Data packageId}
}
\value{
(list) Data package scope, identifier, and revision
}
\description{
Parse package ID into scope, identifier, and revision
}
