% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{parse_quality_report}
\alias{parse_quality_report}
\title{Parse the evaluate quality report}
\usage{
parse_quality_report(qualityReport, full = FALSE)
}
\arguments{
\item{qualityReport}{(xml_document) Evaluate quality report document}

\item{full}{(logical) Return the full report if TRUE, otherwise return only warnings and errors.}
}
\value{
(character) A parsed \code{qualityReport}
}
\description{
Parse the evaluate quality report
}
\examples{


}
