% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_compare.R
\name{pkg_compare}
\alias{pkg_compare}
\title{Compare most data package revisions}
\usage{
pkg_compare(package.id, environment)
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.2').}

\item{environment}{(character) Data repository environment containing the package to be
updated. Can be: 'development', 'staging', 'production'.}
}
\value{
If inconsistencies are found, then a warning is returned.
}
\description{
Compare the contents of the 2 most recent revisions of a data package 
    in the EDI data repository and report differences.
}
\details{
Comparisons:
    \itemize{
        \item{Data entity names - Do data entity names match?}
        \item{Data table columns - Do data tables contain the same number 
        of columns and are column names consistent?}
    }
}
