% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data_descendants.R
\name{pkg_data_descendants}
\alias{pkg_data_descendants}
\title{Get data descendants}
\usage{
pkg_data_descendants(package.id, environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('XMLInternalDocument' 'XMLAbstractDocument') XML containing the the 
    data package ID, data package title, and data package URL.
}
\description{
Get data descendants of a specified parent data package.
}
