% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data_entity_ids.R
\name{pkg_data_entity_ids}
\alias{pkg_data_entity_ids}
\title{Get data entity IDs}
\usage{
pkg_data_entity_ids(package.id, environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(data frame) A data frame with columns:
    \itemize{
        \item{identifier: Data entity identifier}
    }
}
\description{
Get identifiers of data entities.
}
