% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data_entity_names.R
\name{pkg_data_entity_names}
\alias{pkg_data_entity_names}
\title{Get data entity names}
\usage{
pkg_data_entity_names(package.id, environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(data frame) A data frame with columns:
    \itemize{
        \item{}
        \item{}
    }
}
\description{
Get names of data entities and corresponding IDs.
}
