% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data_entity_resource_metadata.R
\name{pkg_data_entity_resource_metadata}
\alias{pkg_data_entity_resource_metadata}
\title{Get data entity resource metadata}
\usage{
pkg_data_entity_resource_metadata(package.id, entity.id, 
    environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{entity.id}{(character) Data entity identifier (e.g. 
"2353ac38985edd6aff140e4c65cb32de")}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('XMLInternalDocument' 'XMLAbstractDocument') Resource metadata.
}
\description{
Get resource metadata for a data entity.
}
