% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data_sources.R
\name{pkg_data_sources}
\alias{pkg_data_sources}
\title{Get data sources of a data package}
\usage{
pkg_data_sources(package.id, environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('XMLInternalDocument' 'XMLAbstractDocument') For each data source, its 
    package identifier, title, and URL values are included (if applicable) 
    as documented in the metadata for the specified data package. Internal 
    data sources include a “packageId” value and a URL to the source 
    metadata. For data sources external to PASTA, the “packageId” element 
    will be empty and a URL value may or not be documented.
}
\description{
Get data sources of a data package. Data sources can be either internal 
    or external to PASTA+.
}
