% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_eml.R
\name{pkg_eml}
\alias{pkg_eml}
\title{Get data package EML}
\usage{
pkg_eml(package.id, environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('XMLInternalDocument' 'XMLAbstractDocument') EML metadata.
}
\description{
Get EML for a specified data package.
}
