% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_evaluate.R
\name{pkg_evaluate}
\alias{pkg_evaluate}
\title{Evaluate data package}
\usage{
pkg_evaluate(path, package.id, environment, user.id, user.pass,
    affiliation)
}
\arguments{
\item{path}{(character) Path to the data package EML file.}

\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.4').}

\item{environment}{(character) Data repository environment to perform the evaluation in.
Can be: 'development', 'staging', 'production'.}

\item{user.id}{(character) Identification of user performing the evaluation.}

\item{user.pass}{(character) Password corresponding with the user.id argument supplied
above.}

\item{affiliation}{(character) Affiliation corresponding with the user.id argument supplied
above. Can be: 'LTER' or 'EDI'.}
}
\description{
Evaluate data package against the Environmental Data Initiative metadata
    quality checker.
}
\note{
Data described by the EML metadata record must be accessible through a
    URL supporting download by the EDI data repository. This URL must be
    supplied for each data entity in the "physical" node of the EML (i.e.
    //physical/distribution/online/url).
}
