% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_ids.R
\name{pkg_ids}
\alias{pkg_ids}
\title{Get data package IDs of a scope}
\usage{
pkg_ids(scope, environment = 'production')
}
\arguments{
\item{scope}{(character) Scope (e.g. 'edi', 'knb-lter-bnz').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Data package identifiers
}
\description{
Get data package IDs for a specific scope.
}
