% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_reserve_id.R
\name{pkg_reserve_id}
\alias{pkg_reserve_id}
\title{Reserve package identifier}
\usage{
pkg_reserve_id(environment, user.id, user.pass, affiliation)
}
\arguments{
\item{scope}{(character) Scope of identifier to be reserved (e.g. edi, knb-lter-ntl).}

\item{environment}{(character) Data repository environment in which to reserve the
identifier. Can be: 'development', 'staging', 'production'.}

\item{user.id}{(character) Identification of user reserving the identifier.}

\item{user.pass}{(character) Password corresponding with the user.id argument supplied
above.}

\item{affiliation}{(character) Affiliation corresponding with the user.id argument supplied
above. Can be: 'LTER' or 'EDI'.}
}
\value{
(character) Package identifier.
}
\description{
Reserve a package identifier in the Environmental Data Initiative
    repository.
}
