% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_revisions.R
\name{pkg_revisions}
\alias{pkg_revisions}
\title{Get revisions of a data package}
\usage{
pkg_revisions(scope, identifier, filter = NULL, 
    environment = 'production')
}
\arguments{
\item{scope}{(character) Data package scope (e.g. 'edi', 'knb-lter-bnz').}

\item{identifier}{(character) Data package identifier (e.g. '100', '275').}

\item{filter}{(character) Filter to data packages, can be 'oldest' or 'newest'.}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(integer) Vector of revisions, if more than one exists, otherwise a 
    single revision number.
}
\description{
Get data package revisions. The request may be filtered by the modifiers 
    “oldest” or “newest”.
}
