% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_update.R
\name{pkg_update}
\alias{pkg_update}
\title{Update data package}
\usage{
pkg_update(path, package.id, environment, user.id, user.pass,
    affiliation)
}
\arguments{
\item{path}{(character) Path to the data package EML file.}

\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.2').}

\item{environment}{(character) Data repository environment containing the package to be
updated. Can be: 'development', 'staging', 'production'.}

\item{user.id}{(character) Identification of user updating the data package.}

\item{user.pass}{(character) Password corresponding with the user.id argument supplied
above.}

\item{affiliation}{(character) Affiliation corresponding with the user.id argument supplied
above. Can be: 'LTER' or 'EDI'.}
}
\description{
Update a data package in the Environmental Data Initiative repository.
}
\note{
Data described by the EML metadata record must be accessible through a
    URL supporting download by the EDI data repository. This URL must be
    supplied for each data entity in the "physical" node of the EML (i.e.
    //physical/distribution/online/url).
}
