% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_staged.R
\name{publish_staged}
\alias{publish_staged}
\title{Publish a "Staged" EDI Data Package}
\usage{
publish_staged(staged_id = NULL, prod_id = NULL, tweet_text = NULL)
}
\arguments{
\item{staged_id}{(character) The full package identifier (scope, accession, revision) of an EDI Data Package in the PASTA staging environment}

\item{prod_id}{(character; optional) The full package identifier (scope, accession, revision) of a reserved EDI Data Package that the stage package should be}

\item{tweet_text}{(character; optional) The text for a tweet about the recent data publication. If provided, a URL for a pre-populated tweet (with EDI and the data package both tagged) is returned.}
}
\value{
transaction (character) Transaction identifier. May be used in a subsequent call to:
\itemize{
  \item \code{check_status_create()} to determine the operation status
  \item \code{read_data_package()} to obtain the data package resource map
}
}
\description{
This function accesses a data package from the EDI Staging Environment, assigns a new identifier, and publishes to the EDI Data Portal.
}
\details{
User can provide their own identifier or else an identifier will be automatically reserved and assigned.
}
\examples{

\dontrun{

# Upload staged package edi.101.1 to the next available edi-scoped, production id:

publish_staged('edi.101.1')

# Upload staged packaged edi.101.1 to a user-specified id:

publish_staged('edi.101.1', 'knb-lter-ntl.5.22')

# Upload staged package edi.101.1 and generate a URL for a pre-populated tweet:
 
 publish_staged('edi.101.1', tweet_text = "See our recent data package publication!")

}

}
