% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity_acl.R
\name{read_data_entity_acl}
\alias{read_data_entity_acl}
\title{Read data entity access control list}
\usage{
read_data_entity_acl(packageId, entityId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{entityId}{(character) Data entity identifier}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) The access control list for the data entity. Please note: only a very limited set of users are authorized to use this service method.
}
\description{
Read data entity access control list
}
\note{
User authentication is required (see \code{login()})
}
\examples{

}
