% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_acl.R
\name{read_data_package_acl}
\alias{read_data_package_acl}
\title{Read data package Access Control List}
\usage{
read_data_package_acl(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) Access Control List (ACL) for \code{packageId}. Please note: only a very limited set of users are authorized to use this service method.
}
\description{
Read data package Access Control List
}
\examples{


}
