% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_report_resource_metadata.R
\name{read_data_package_report_resource_metadata}
\alias{read_data_package_report_resource_metadata}
\title{Read data package report resource metadata}
\usage{
read_data_package_report_resource_metadata(
  packageId,
  as = "data.frame",
  env = "production"
)
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{as}{(character) Format of the returned object. Can be: "data.frame" 
or "xml".}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(data.frame or xml_document) Report resource metadata
}
\description{
Read data package report resource metadata
}
\examples{
\dontrun{

# Read resource metadata
resourceMetadata <- read_data_package_report_resource_metadata(
  packageId = "knb-lter-mcm.9129.3"
)
}
}
\seealso{
Other Accessing:
\code{\link[=read_data_entity]{read_data_entity()}},
\code{\link[=read_data_entity_checksum]{read_data_entity_checksum()}},
\code{\link[=read_data_entity_name]{read_data_entity_name()}},
\code{\link[=read_data_entity_names]{read_data_entity_names()}},
\code{\link[=read_data_entity_resource_metadata]{read_data_entity_resource_metadata()}},
\code{\link[=read_data_entity_size]{read_data_entity_size()}},
\code{\link[=read_data_entity_sizes]{read_data_entity_sizes()}},
\code{\link[=read_data_package]{read_data_package()}},
\code{\link[=read_data_package_archive]{read_data_package_archive()}},
\code{\link[=read_data_package_citation]{read_data_package_citation()}},
\code{\link[=read_data_package_doi]{read_data_package_doi()}},
\code{\link[=read_data_package_error]{read_data_package_error()}},
\code{\link[=read_data_package_from_doi]{read_data_package_from_doi()}},
\code{\link[=read_data_package_report]{read_data_package_report()}},
\code{\link[=read_data_package_report_checksum]{read_data_package_report_checksum()}},
\code{\link[=read_data_package_report_summary]{read_data_package_report_summary()}},
\code{\link[=read_data_package_resource_metadata]{read_data_package_resource_metadata()}},
\code{\link[=read_evaluate_report]{read_evaluate_report()}},
\code{\link[=read_evaluate_report_summary]{read_evaluate_report_summary()}},
\code{\link[=read_metadata]{read_metadata()}},
\code{\link[=read_metadata_checksum]{read_metadata_checksum()}},
\code{\link[=read_metadata_dublin_core]{read_metadata_dublin_core()}},
\code{\link[=read_metadata_entity]{read_metadata_entity()}},
\code{\link[=read_metadata_format]{read_metadata_format()}},
\code{\link[=read_metadata_resource_metadata]{read_metadata_resource_metadata()}}
}
\concept{Accessing}
