% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata_acl.R
\name{read_metadata_acl}
\alias{read_metadata_acl}
\title{Read metadata access control list}
\usage{
read_metadata_acl(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) Metadata access control list. Please note: only a very limited set of users are authorized to use this service method.
}
\description{
Read metadata access control list
}
\note{
User authentication is required (see \code{login()})
}
\examples{

}
