% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_value_codes.R
\name{replace_missing_value_codes}
\alias{replace_missing_value_codes}
\title{Replace missing value codes with a new value}
\usage{
replace_missing_value_codes(
  x = NULL,
  eml = NULL,
  with = NA,
  file = NULL,
  coerce = TRUE
)
}
\arguments{
\item{x}{(data frame) Data with missing value codes to convert to NA.}

\item{eml}{(xml_document xml_node) EML metadata listing missing value codes for 
\code{x}. Use \code{EDIutils::read_metadata()} or 
\code{xml2::read_xml()} to read the EML file.}

\item{with}{(character or NA) New missing value code. This value will replace those 
listed in \code{x} as specified by \code{eml}. Default is NA.}

\item{file}{(character) Name of the file, which \code{x} was created from (e.g. 
"Resin_Deposition.csv". Use this argument if a column in \code{x} 
has more than one match in \code{x} (see details).}

\item{coerce}{(logical) Columns of \code{x} defined as numeric in \code{eml} will be
coerced to numeric after missing value code replacement. Default is 
TRUE.}
}
\value{
(data frame) \code{x} with missing value codes replaced with 
    \code{with}.
}
\description{
Missing value codes vary across data packages and some computational 
    environments have preferences (e.g. R uses NA). This function transform 
    these codes to make downstream use simpler.
}
\details{
For each column in \code{x} a matching attribute is searched for in 
    \code{eml}. If one is found, then the missing value code listed in 
    \code{eml} is replaced in the data with \code{with}. If more than one
    attribute is found then an error is returned prompting the use of the 
    \code{file} argument. \code{file} maps to 
    /eml/dataset/dataTable/physical/objectName in \code{eml} and creates an 
    unambiguous column name/attribute match.
    
    Columns in \code{x} that are defined in \code{eml} as numeric, but 
    are read in as character due to the missing value code, are converted
    to numeric.
}
