% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_user_agent}
\alias{set_user_agent}
\title{Set EDIutils user agent for http requests}
\usage{
set_user_agent()
}
\value{
(request) EDIutils user agent
}
\description{
Set EDIutils user agent for http requests
}
