% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{skip_if_logged_out}
\alias{skip_if_logged_out}
\title{Skip tests when logged out}
\usage{
skip_if_logged_out()
}
\description{
Skip tests when logged out
}
\details{
Facilitates testing of functions requiring authentication
}
