% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{skip_if_missing_eml_config}
\alias{skip_if_missing_eml_config}
\title{Skip tests when EML configuration is missing}
\usage{
skip_if_missing_eml_config()
}
\description{
Skip tests when EML configuration is missing
}
\details{
Facilitates testing create, update, and delete for a test data package
}
