% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{summarize_evaluate_report}
\alias{summarize_evaluate_report}
\title{Summarize the evaluate quality report}
\usage{
summarize_evaluate_report(
  transaction,
  with_exceptions = TRUE,
  tier = "production"
)
}
\arguments{
\item{transaction}{(character) Transaction identifier}

\item{with_exceptions}{(logical) Convert quality report warnings and errors to R warnings and errors}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
(message/warning/error) A message listing the total number of checks resulting in valid, info, warn, and error status. Exceptions are raised if warnings and errors are found and \code{with_exceptions} is TRUE.
}
\description{
Summarize the evaluate quality report
}
\details{
Get \code{transaction} from \code{evaluate_data_package()}
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{
path <- "/Users/me/Documents/edi.468.1.xml"
transaction <- evaluate_data_package(path)
summarize_evaluate_report(transaction)
}

}
