% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{summarize_evalutation_report}
\alias{summarize_evalutation_report}
\title{Summarize evaluation report}
\usage{
summarize_evalutation_report(eval_report)
}
\arguments{
\item{eval_report}{(character) Evaluation report XML as a character string}
}
\value{
(character) Summary of evaluation report
}
\description{
Summarize evaluation report
}
