% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_tweet.R
\name{template_tweet}
\alias{template_tweet}
\title{Generate a URL for a pre-populated tweet}
\usage{
template_tweet(packageId, text)
}
\arguments{
\item{packageId}{(character) The full package identifier (scope, accession, revision) of an EDI Data Package in the PASTA production environment}

\item{text}{(character) The text for the tweet to be generated}
}
\value{
A URL for a pre-populated tweet that links to a data package and tags EDI's twitter account
}
\description{
Generate a URL for a pre-populated tweet
}
\examples{

template_tweet('edi.101.1', "See our recent data package publication!")

}
