% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{text2char}
\alias{text2char}
\title{Convert newline separated text to character vector}
\usage{
text2char(txt)
}
\arguments{
\item{txt}{(character) New line separated character string returned from \code{httr::content(resp, as = "text", encoding = "UTF-8")}}
}
\value{
(character) \code{txt} converted to character vector
}
\description{
Convert newline separated text to character vector
}
