% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish_staged.R
\name{transfer_journal_citations}
\alias{transfer_journal_citations}
\title{Transfer linked journal citations from a staged data package to a published data package}
\usage{
transfer_journal_citations(staged_id, new_id, all_versions = T)
}
\arguments{
\item{staged_id}{(character) The full package identifier (scope, accession, revision) of an EDI Data Package in the PASTA staging environment}

\item{all_versions}{(boolean) If true, checks all revisions of the staged package. If false, only checks the specified revision}

\item{prod_id}{(character; optional) The full package identifier (scope, accession, revision) of a reserved EDI Data Package that the stage package should be}
}
\value{
The linked journal citations from the staged data package are now linked to the published data package
}
\description{
Transfer linked journal citations from a staged data package to a published data package
}
\examples{

# Transfer journal citations from staged package edi.101.1 to published data package edi.787.1

transfer_journal_citations('edi.101.1', 'edi.787.1', all_versions = T)

# Transfer journal citations from all versions of staged package edi.101 to published data package edi.787.1

transfer_journal_citations('edi.101.6', 'edi.787.1', all_versions = T)

}
