% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_arguments}
\alias{validate_arguments}
\title{Validate function arguments}
\usage{
validate_arguments(x)
}
\arguments{
\item{x}{(list) A named list of arguments and corresponding values.}
}
\description{
Validate the contents of function arguments.
}
