% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{xml2df}
\alias{xml2df}
\title{Convert minimally nested XML to data.frame}
\usage{
xml2df(xml)
}
\arguments{
\item{xml}{(xml_document xml_node) XML with one level of nesting}
}
\value{
(data.frame) A data.frame of \code{xml}
}
\description{
Convert minimally nested XML to data.frame
}
