% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_read_data_entity_names.R
\name{api_read_data_entity_names}
\alias{api_read_data_entity_names}
\title{Read data entity names}
\usage{
api_read_data_entity_names(package.id, identifier, 
    environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Data entity names
}
\description{
Read Data Entity Names operation, specifying the scope, identifier, 
    and revision of the data package whose data entity names are to be 
    read in the URI. Only data entities that the user is authorized to 
    read are included in the list.
}
