% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_status_create.R
\name{check_status_create}
\alias{check_status_create}
\title{Check data package creation status}
\usage{
check_status_create(transaction, packageId, wait = TRUE, tier = "production")
}
\arguments{
\item{transaction}{(character) Transaction identifier}

\item{packageId}{(character) Data package identifier}

\item{wait}{(logical) Wait for evaluation to complete? See details below.}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(logical) TRUE if creation has completed, FALSE if in progress, and error if an error was encountered while processing the request
}
\description{
Check data package creation status
}
\details{
If \code{wait = TRUE}, then the function will enter a "while" loop checking every 2 seconds for the completed evaluation report. If \code{wait = FALSE}, then the function will only check once and return the result.
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{
path <- "/Users/me/Documents/edi.468.1.xml"
transaction <- create_data_package(path)
packageId <- "edi.468.1"
check_status_create(transaction, packageId)
}

}
