% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_status_evaluate.R
\name{check_status_evaluate}
\alias{check_status_evaluate}
\title{Check status of data package evaluation}
\usage{
check_status_evaluate(transaction, tier = "production")
}
\arguments{
\item{transaction}{(character) Transaction identifier}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(logical) TRUE if evaluation has completed, FALSE if in progress, and error if an error was encountered while processing the request
}
\description{
Check status of data package evaluation
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{
path <- "/Users/me/Documents/edi.468.1.xml"
transaction <- evaluate_data_package(path)
check_status_evaluate(transaction)
}

}
