% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_status_evaluate.R
\name{check_status_evaluate}
\alias{check_status_evaluate}
\title{Check status of data package evaluation}
\usage{
check_status_evaluate(transaction, wait = TRUE, env = "production")
}
\arguments{
\item{transaction}{(character) Transaction identifier}

\item{wait}{(logical) Wait for evaluation to complete? See details below.}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(logical) TRUE if evaluation has completed, FALSE if in progress, and error if an error was encountered while processing the request
}
\description{
Check status of data package evaluation
}
\details{
If \code{wait = TRUE}, then the function will enter a "while" loop checking every 2 seconds for the completed evaluation report. If \code{wait = FALSE}, then the function will only check once and return the result.
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{
login()

# Evaluate data package
eml <- "./data/edi.595.1.xml"
transaction <- evaluate_data_package(eml, env = "staging")

# Check evaluation status
status <- check_status_evaluate(transaction, env = "staging")
status
#> [1] TRUE

logout()
}

}
