% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_eml.R
\name{compare_eml}
\alias{compare_eml}
\title{Are there meaningful differences between EML documents?}
\usage{
compare_eml(newest, previous, return.all = FALSE)
}
\arguments{
\item{newest}{(xml_document, xml_node) EML of the newest version of a data package, where inputs are returned from \code{api_read_metadata()}.}

\item{previous}{(xml_document, xml_node) EML of the previous version of a data package, where inputs are returned from \code{api_read_metadata()}.}

\item{return.all}{(logical) Return all differences? Default is FALSE, i.e. only return meaningful differences. Meaningful differences do not include elements expected to change between versions (e.g. number of rows, file size, temporal coverage).}
}
\value{
(character) XPaths of nodes that differ between versions
}
\description{
For discovering changes within a dataset that may have affect downstream processes relying on consistent dataset structure and meaning. This is useful in workflow automation where reporting such changes can expedite trouble shooting and manual intervention.
}
\details{
XPaths of checked nodes (and whether "meaningful"):
\itemize{
  \item{.//dataset/abstract (TRUE)}
  \item{.//dataset/coverage/geographicCoverage (FALSE)}
  \item{.//dataset/coverage/temporalCoverage (FALSE)}
  \item{.//dataset/coverage/taxonomicCoverage (FALSE)}
  \item{.//dataset/keywordSet (FALSE)}
  \item{.//dataTable/physical/objectName (TRUE)}
  \item{.//dataTable/physical/size (FALSE)}
  \item{.//dataTable/physical/authentication (FALSE)}
  \item{.//dataTable/physical/dataFormat/textFormat/numHeaderLines (TRUE)}
  \item{.//dataTable/physical/dataFormat/textFormat/recordDelimiter (FALSE)}
  \item{.//dataTable/physical/dataFormat/textFormat/attributeOrientation (TRUE)}
  \item{.//dataTable/physical/dataFormat/textFormat/simpleDelimited/fieldDelimiter (TRUE)}
  \item{.//dataTable/attributeList (TRUE)}
  \item{.//dataTable/numberOfRecords (FALSE)}
  \item{.//otherEntity/physical/objectName (TRUE)}
  \item{.//otherEntity/physical/size (FALSE)}
  \item{.//otherEntity/physical/authentication (FALSE)}
  \item{.//otherEntity/physical/dataFormat/textFormat/numHeaderLines (TRUE)}
  \item{.//otherEntity/physical/dataFormat/textFormat/recordDelimiter (TRUE)}
  \item{.//otherEntity/physical/dataFormat/textFormat/attributeOrientation (TRUE)}
  \item{.//otherEntity/physical/dataFormat/textFormat/simpleDelimited/fieldDelimiter (TRUE)}
  \item{.//otherEntity/attributeList (TRUE)}
}
}
\examples{

# Return only "meaningful" differences (default behavior)
compare_eml(
  newest = api_read_metadata("knb-lter-hfr.118.32"),
  previous = api_read_metadata("knb-lter-hfr.118.31"))
  
# Return all differences
compare_eml(
  newest = api_read_metadata("knb-lter-hfr.118.32"),
  previous = api_read_metadata("knb-lter-hfr.118.31"),
  return.all = TRUE)

}
