% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{construct_dn}
\alias{construct_dn}
\title{Construct a users distinguished name}
\usage{
construct_dn(userId, ou)
}
\arguments{
\item{userId}{(character) PASTA userId}

\item{ou}{(character) Organizational unit in which \code{userId} belongs. Can be "EDI" or "LTER".}
}
\value{
(character) Distinguished name
}
\description{
Construct a users distinguished name
}
\examples{
get_distinguished_name("csmith", "EDI")

}
