% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_package.R
\name{create_data_package}
\alias{create_data_package}
\title{Create data package}
\usage{
create_data_package(
  path,
  package.id,
  environment,
  user.id,
  user.pass,
  affiliation
)
}
\arguments{
\item{path}{(character) Path to the data package EML file.}

\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}

\item{user.id}{(character) Identification of user creating the data package.}

\item{user.pass}{(character) Password corresponding with the user.id argument supplied
above.}

\item{affiliation}{(character) Affiliation corresponding with the user.id argument supplied
above. Can be: 'LTER' or 'EDI'.}
}
\description{
Creates a new PASTA data package by providing the EML document describing the data package to be created in the request message body and returning a transaction identifier in the response message body as plain text; the transaction identifier may be used in a subsequent call to Read Data Package Error to determine the operation status; see Read Data Package to obtain the data package resource map if the operation completed successfully.
}
\details{
POST : https://pasta.lternet.edu/package/eml
}
\examples{
Upload a new data package where knb-lter-lno.1.1.xml is the filesystem EML document in XML format

}
