% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_package.R
\name{create_data_package}
\alias{create_data_package}
\title{Create data package}
\usage{
create_data_package(eml, env = "production")
}
\arguments{
\item{eml}{(character) Full path to an EML file describing the data package to be created}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
transaction (character) Transaction identifier. May be used in a subsequent call to:
\itemize{
  \item \code{check_status_create()} to determine the operation status
  \item \code{read_data_package()} to obtain the data package resource map
}
}
\description{
Create data package
}
\details{
Each data entity described in \code{eml} must be accompanied by a web accessible URL at the EML XPath ".//physical/distribution/online/url". The EDI data repository downloads the data entities via this URL. The URLs must be static and not have any redirects otherwise the data entities will not be downloaded.
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{
login()

# Create data package
eml <- "./data/edi.595.1.xml"
transaction <- create_data_package(eml, env = "staging")

# Check creation status
packageId <- "edi.595.1"
status <- check_status_create(transaction, packageId, env = "staging")
status
#> [1] TRUE

logout()
}

}
