% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_package.R
\name{create_data_package}
\alias{create_data_package}
\title{Create data package}
\usage{
create_data_package(eml, tier = "production")
}
\arguments{
\item{eml}{(character) Full path to an EML file}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
transaction (character) Transaction identifier. Use this value with: 
\itemize{
  \item \code{check_status_create()} to see if creation has completed or if any errors occurred while processing the request
  \item \code{read_evaluate_report()} to read the evaluation report
  \item \code{read_data_package()} to obtain the data package resource map
}
}
\description{
Create data package
}
\details{
Each data entity described in \code{eml} must be accessible to the EDI Data Repository through URLs listed in the \code{eml} at the XPath ".//physical/distribution/online/url". This URL should be static and not present the EDI Data Repository with any redirects, otherwise the data entity will not be accessible.
}
\examples{
\dontrun{
path <- "/Users/me/Documents/edi.468.1.xml"
create_data_package(path)
}

}
