% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_package_archive.R
\name{create_data_package_archive}
\alias{create_data_package_archive}
\title{Create data package archive (zip)}
\usage{
create_data_package_archive(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(character) A transaction identifier
}
\description{
Create data package archive (zip)
}
\details{
The transaction identifier may be used in a subsequent call to \code{read_data_package_error()} to determine the operation status or to \code{read_data_package_archive()} to obtain the Zip archive.
}
\examples{
create_data_package_archive("knb-lter-sev.31999.1")

}
