% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_package_archive.R
\name{create_data_package_archive}
\alias{create_data_package_archive}
\title{Create data package archive (zip)}
\usage{
create_data_package_archive(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
transaction (character) Transaction identifier. May be used in a subsequent call to:
\itemize{
  \item \code{read_data_package_error()} to determine the operation status
  \item \code{read_data_package_archive()} to obtain the Zip archive
}
}
\description{
Create data package archive (zip)
}
\examples{
create_data_package_archive("knb-lter-sev.31999.1")

}
