% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_package_archive.R
\name{create_data_package_archive}
\alias{create_data_package_archive}
\title{Create data package archive (zip)}
\usage{
create_data_package_archive(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
transaction (character) Transaction identifier.
}
\description{
This function is DEPRECATED.
}
\seealso{
Other Miscellaneous: 
\code{\link{create_dn}()},
\code{\link{is_authorized}()}
}
\concept{Miscellaneous}
