% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_dn}
\alias{create_dn}
\title{Construct a users distinguished name}
\usage{
create_dn(userId, ou = "EDI")
}
\arguments{
\item{userId}{(character) PASTA userId}

\item{ou}{(character) Organizational unit in which \code{userId} belongs. Can be "EDI" or "LTER". All \code{userId} issued after "2020-05-01" have \code{ou = "EDI"}.}
}
\value{
(character) Distinguished name
}
\description{
Construct a users distinguished name
}
\examples{
create_dn("csmith")

}
