% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_event_subscription.R
\name{create_event_subscription}
\alias{create_event_subscription}
\title{Create event subscription}
\usage{
create_event_subscription(packageId, url, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{url}{(character) Where the event notification will be sent}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(numeric) Event subscription ID
}
\description{
Create event subscription
}
\note{
User authentication is required (see \code{login()})

 The \code{url} must have "http" as its scheme and must be able to receive POST requests with MIME type text/plain. Additionally, because the \code{url} will be passed in an XML body, some characters must be escaped, such as ampersands from & to &amp;.
}
\examples{
\dontrun{
packageId <- "knb-lter-vcr.340.1"
url <- "https://some.server.org"
create_event_subscription(packageId, url)
}

}
