% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_event_subscription.R
\name{create_event_subscription}
\alias{create_event_subscription}
\title{Create event subscription}
\usage{
create_event_subscription(
  path,
  package.id,
  environment,
  user.id,
  user.pass,
  affiliation
)
}
\arguments{
\item{path}{(character) Path to the data package EML file.}

\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}

\item{user.id}{(character) Identification of user creating the data package.}

\item{user.pass}{(character) Password corresponding with the user.id argument supplied
above.}

\item{affiliation}{(character) Affiliation corresponding with the user.id argument supplied
above. Can be: 'LTER' or 'EDI'.}
}
\description{
Create Event Subscription creates a new event subscription.
}
\details{
POST : https://pasta.lternet.edu/package/event/eml
}
\examples{

}
