% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_event_subscription.R
\name{create_event_subscription}
\alias{create_event_subscription}
\title{Create event subscription}
\usage{
create_event_subscription(tier = "production")
}
\arguments{
\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
A new event subscription
}
\description{
Create event subscription
}
\examples{

}
