% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_journal_citation.R
\name{create_journal_citation}
\alias{create_journal_citation}
\title{Create journal citation}
\usage{
create_journal_citation(
  title,
  creator,
  online.description,
  url,
  contact,
  path = NULL,
  file.name = NULL
)
}
\arguments{
\item{title}{(character) Title of external data source.}

\item{creator}{(list) Named list of EML creator elements and values. Supprted elements 
are: givenName, surName, organizationName, electronicMailAddress}

\item{online.description}{(character) Description of online resource.}

\item{url}{(character) URL of online resource.}

\item{contact}{(list) Named list of EML contact elements and values. Supprted elements 
are: givenName, surName, organizationName, electronicMailAddress}

\item{path}{(character) Where the XML will be written.}

\item{file.name}{(character) Name of file to be written.}
}
\value{
('xml_document' 'xml_node') EML provenance metadata
}
\description{
Create Journal Citation operation, creates a new journal citation entry in PASTA. An XML document containing metadata for the journal citation must be supplied in the HTTP request body.
}
\details{
POST : https://pasta.lternet.edu/package/citation/eml
}
\examples{
# Using curl to create a journal citation with the XML metadata stored in a file:

}
