% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_journal_citation.R
\name{create_journal_citation}
\alias{create_journal_citation}
\title{Create journal citation}
\usage{
create_journal_citation(
  packageId,
  articleDoi = NULL,
  articleUrl = NULL,
  articleTitle = NULL,
  journalTitle = NULL,
  env = "production"
)
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{articleDoi}{(character) Article DOI. Required if \code{articleUrl} is missing.}

\item{articleUrl}{(character) Article URL. Required if \code{articleDoi} is missing.}

\item{articleTitle}{(character) Article title}

\item{journalTitle}{(character) Journal title}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(numeric) Journal citation identifier
}
\description{
Create journal citation
}
\details{
Creates a new journal citation entry in the EDI data repository
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{
res <- create_journal_citation(
  packageId = "edi.17.1", 
  articleDoi = "https://doi.org/10.1890/11-1026.1",
  articleTitle = "Corridors promote fire via connectivity and edge effects",
  journalTitle = "Ecological Applications")
}

}
