% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reservation.R
\name{create_reservation}
\alias{create_reservation}
\title{Create reservation}
\usage{
create_reservation(scope, tier = "production")
}
\arguments{
\item{scope}{(character) Scope of data package}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
(numeric) Identifier of reserved data package (i.e. the second component of a \code{packageId})
}
\description{
Create reservation
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{
login()
create_reservation("edi")
}

}
