% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_data_package.R
\name{delete_data_package}
\alias{delete_data_package}
\title{Delete data package}
\usage{
delete_data_package(scope, identifier, tier = "production")
}
\arguments{
\item{scope}{(character) Scope of data package}

\item{identifier}{(numeric) Identifier of data package}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
(logical) TRUE if deleted
}
\description{
Delete data package
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{

}

}
