% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_event_subscription.R
\name{delete_event_subscription}
\alias{delete_event_subscription}
\title{Delete event subscription}
\usage{
delete_event_subscription(
  path,
  package.id,
  environment,
  user.id,
  user.pass,
  affiliation
)
}
\arguments{
\item{path}{(character) Path to the data package EML file.}

\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}

\item{user.id}{(character) Identification of user creating the data package.}

\item{user.pass}{(character) Password corresponding with the user.id argument supplied
above.}

\item{affiliation}{(character) Affiliation corresponding with the user.id argument supplied
above. Can be: 'LTER' or 'EDI'.}
}
\description{
Delete Event Subscription deletes the event subscription with the specified ID from the subscription database. After “deletion,” the subscription might still exist in the subscription database, but it will be inactive - it will not conflict with future creation requests, it cannot be read, and it will not be notified of events.
}
\details{
DELETE : https://pasta.lternet.edu/package/event/eml/{subscriptionId}
}
\examples{

}
