% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_journal_citation.R
\name{delete_journal_citation}
\alias{delete_journal_citation}
\title{Delete journal citation}
\usage{
delete_journal_citation(journalCitationId, env = "production")
}
\arguments{
\item{journalCitationId}{(numeric) Journal citation identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(logical) TRUE if deleted
}
\description{
Delete journal citation
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{
journalCitationId <- 65
delete_journal_citation(journalCitationId)
}

}
