% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_journal_citation.R
\name{delete_journal_citation}
\alias{delete_journal_citation}
\title{Delete journal citation}
\usage{
delete_journal_citation(
  path,
  package.id,
  environment,
  user.id,
  user.pass,
  affiliation
)
}
\arguments{
\item{path}{(character) Path to the data package EML file.}

\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}

\item{user.id}{(character) Identification of user creating the data package.}

\item{user.pass}{(character) Password corresponding with the user.id argument supplied
above.}

\item{affiliation}{(character) Affiliation corresponding with the user.id argument supplied
above. Can be: 'LTER' or 'EDI'.}
}
\description{
Delete Journal Citation operation, deletes the journal citation entry with the specified ID from the journal citation table. Requires authentication by the owner of the journal citation entry.
}
\details{
DELETE : https://pasta.lternet.edu/package/citation/eml/{journalCitationId}
}
\examples{
# Using curl to delete the journal citation with identifier value 15, owned by user “ucarroll”:

}
