% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_reservation.R
\name{delete_reservation}
\alias{delete_reservation}
\title{Delete reservation}
\usage{
delete_reservation(scope, environment, user.id, user.pass, affiliation)
}
\arguments{
\item{scope}{(character) Scope of identifier to be reserved (e.g. edi, knb-lter-ntl).}

\item{environment}{(character) Data repository environment in which to reserve the
identifier. Can be: 'development', 'staging', 'production'.}

\item{user.id}{(character) Identification of user reserving the identifier.}

\item{user.pass}{(character) Password corresponding with the user.id argument supplied
above.}

\item{affiliation}{(character) Affiliation corresponding with the user.id argument supplied
above. Can be: 'LTER' or 'EDI'.}
}
\value{
(character) Package identifier.
}
\description{
Delete Reservation operation, deletes an existing reservation from PASTA. The same user who originally authenticated to create the reservation must authenticate to delete it, otherwise a “401 Unauthorized” response is returned. When successfully deleted, a “200 OK” response is returned, and the integer value of the deleted reservation identifier value is returned in the web service response body.
}
\details{
DELETE : https://pasta.lternet.edu/package/reservations/eml/{scope}/{identifier}
}
\examples{
# Using curl to delete an existing reservation for scope (“edi”) and identifier (“12”):

}
