% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_reservation.R
\name{delete_reservation}
\alias{delete_reservation}
\title{Delete reservation}
\usage{
delete_reservation(scope, identifier, env = "production")
}
\arguments{
\item{scope}{(character) Scope of data package}

\item{identifier}{(numeric) Identifier of data package}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(numeric) The deleted reservation identifier value
}
\description{
Delete reservation
}
\note{
User authentication is required (see \code{login()}). The same user who originally authenticated to create the reservation must authenticate to delete it.
}
\examples{
\dontrun{
login()
identifier <- create_reservation("edi")
delete_reservation("edi", identifier)
}

}
