% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_event_subscription.R
\name{execute_event_subscription}
\alias{execute_event_subscription}
\title{Execute event subscription}
\usage{
execute_event_subscription(scope, environment, user.id, user.pass, affiliation)
}
\arguments{
\item{scope}{(character) Scope of identifier to be reserved (e.g. edi, knb-lter-ntl).}

\item{environment}{(character) Data repository environment in which to reserve the
identifier. Can be: 'development', 'staging', 'production'.}

\item{user.id}{(character) Identification of user reserving the identifier.}

\item{user.pass}{(character) Password corresponding with the user.id argument supplied
above.}

\item{affiliation}{(character) Affiliation corresponding with the user.id argument supplied
above. Can be: 'LTER' or 'EDI'.}
}
\value{
(character) Package identifier.
}
\description{
Execute Event Subscription operation, specifying the ID of the event subscription whose URL is to be executed. Used to execute a particular subscription in the event manager, via an HTTP POST request. Upon notification, the event manager queries its database for the subscription matching the specified subscriptionId. POST requests are then made (asynchronously) to the matching subscription.

The request headers must contain an authorization token. If the request is successful, an HTTP response with status code 200 ‘OK’ is returned. If the request is unauthorized, based on the content of the authorization token and the current access control rule for event notification, status code 401 ‘Unauthorized’ is returned. If the request contains an error, status code 400 ‘Bad Request’ is returned, with a description of the encountered error.
}
\details{
POST : https://pasta.lternet.edu/package/event/eml/{subscriptionId}
}
\examples{

}
