% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_event_subscription.R
\name{execute_event_subscription}
\alias{execute_event_subscription}
\title{Execute event subscription}
\usage{
execute_event_subscription(subscriptionId, tier = "production")
}
\arguments{
\item{subscriptionId}{(numeric) Event subscription identifier}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
(logical) TRUE if the event subscription was executed
}
\description{
Execute event subscription
}
\details{
Upon notification, the event manager queries its database for the subscription matching the specified subscriptionId. POST requests are then made (asynchronously) to the matching subscription.
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{
packageId <- "knb-lter-vcr.340.1"
url <- "https://some.server.org"
subscriptionId <- create_event_subscription(packageId, url)
execute_event_subscription(subscriptionId)
}

}
