% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_audit_count.R
\name{get_audit_count}
\alias{get_audit_count}
\title{Get audit count}
\usage{
get_audit_count(attr.name, package.id, environment = "production")
}
\arguments{
\item{attr.name}{(character) Attribute name}

\item{package.id}{(character) Data package identifier (e.g. knb-lter-cap.627.3) in the 
EDI Data Repository.}

\item{evironment}{(character) Repository environment in which the L0 and L1 exist. Some repositories have development, staging, and production environments which are distinct from one another. This argument allows execution for the \code{update_L1} workflow within the context of one of these environments. Default is "production".}
}
\value{
(list) A named vector with these attribute elements:
    \itemize{
        \item{name} <attributeName>
        \item{definition} <attributeDefinition> 
        \item{unit} <standardUnit> or <customUnit>
    }
}
\description{
Returns a count of the number of audit records matching the query parameters as specified in the request.
}
\details{
GET : https://pasta.lternet.edu/audit/count
}
\examples{

}
