% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_audit_record.R
\name{get_audit_record}
\alias{get_audit_record}
\title{Get audit record}
\usage{
get_audit_record(oid, env = "production")
}
\arguments{
\item{oid}{(numeric) Audit identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) An audit record
}
\description{
Get audit record
}
\note{
User authentication is required (see \code{login()})
}
\examples{
query <- "serviceMethod=createDataPackage&limit=5"
recent_uploads <- get_recent_uploads(query)
oid <- xml2::xml_text(xml2::xml_find_all(recent_uploads, ".//oid"))[1]
get_audit_record(oid)

}
