% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_docid_reads.R
\name{get_docid_reads}
\alias{get_docid_reads}
\title{Get doc ID reads}
\usage{
get_docid_reads(scope, identifier, tier = "production")
}
\arguments{
\item{scope}{(character) Scope of data package (i.e. the first component of a \code{packageId})}

\item{identifier}{(numeric) Identifier of data package (i.e. the second component of a \code{packageId})}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) Summary of all the successful reads (total reads and non-robot reads) for all the resources of a given \code{scope} and \code{identifier}.
}
\description{
Get doc ID reads
}
\examples{
get_docid_reads("knb-lter-sgs", "817")

}
