% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eml_attribute.R
\name{get_eml_attribute}
\alias{get_eml_attribute}
\title{Get metadata for an attribute}
\usage{
get_eml_attribute(attr.name, package.id)
}
\arguments{
\item{attr.name}{(character) Attribute name}

\item{package.id}{(character) Data package identifier (e.g. knb-lter-cap.627.3) in the 
EDI Data Repository.}
}
\value{
(list) A named vector with these attribute elements:
    \itemize{
        \item{name} <attributeName>
        \item{definition} <attributeDefinition> 
        \item{unit} <standardUnit> or <customUnit>
    }
}
\description{
Get metadata for an attribute.
}
