% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_subscription.R
\name{get_event_subscription}
\alias{get_event_subscription}
\title{Get event subscription}
\usage{
get_event_subscription(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
("xml_document" "xml_node") EML metadata.
}
\description{
Get Event Subscription returns the event subscription with the specified ID.
}
\details{
GET : https://pasta.lternet.edu/package/event/eml/{subscriptionId}
}
\examples{

}
