% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_subscription.R
\name{get_event_subscription}
\alias{get_event_subscription}
\title{Get event subscription}
\usage{
get_event_subscription(subscriptionId, env = "production")
}
\arguments{
\item{subscriptionId}{(numeric) Event subscription identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Subscription metadata
}
\description{
Get event subscription
}
\note{
User authentication is required (see \code{login()})
}
\examples{
packageId <- "knb-lter-vcr.340.1"
url <- "https://some.server.org"
subscriptionId <- create_event_subscription(packageId, url)
get_event_subscription(subscriptionId)

}
