% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_subscription.R
\name{get_event_subscription}
\alias{get_event_subscription}
\title{Get event subscription}
\usage{
get_event_subscription(subscriptionId, env = "production")
}
\arguments{
\item{subscriptionId}{(numeric) Event subscription identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Subscription metadata
}
\description{
Get event subscription
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{

login()

# Get subscription
subscription <- get_event_subscription(
  subscriptionId = 21, 
  env = "staging")
subscription
#> {xml_document}
#> <subscriptions>
#> [1] <subscription type="eml">\n  <id>21</id>\n  <creator>uid=csmith, ...

xml2::xml_find_first(subscriptions, "subscription")
#> {xml_node}
#> <subscription type="eml">
#> [1] <id>21</id>
#> [2] <creator>uid=csmith,o=EDI,dc=edirepository,dc=org</creator>
#> [3] <packageId>edi.94</packageId>
#> [4] <url>https://regan.edirepository.org/ecocom-listener</url>

logout()
}

}
