% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_subscription.R
\name{get_event_subscription}
\alias{get_event_subscription}
\title{Get event subscription}
\usage{
get_event_subscription(subscriptionId, tier = "production")
}
\arguments{
\item{subscriptionId}{(numeric) Event subscription identifier}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Subscription metadata
}
\description{
Get Event Subscription returns the event subscription with the specified ID.
}
\examples{
packageId <- "knb-lter-vcr.340.1"
url <- "https://some.server.org"
subscriptionId <- create_event_subscription(packageId, url)
get_event_subscription(subscriptionId)

}
