% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_subscription_schema.R
\name{get_event_subscription_schema}
\alias{get_event_subscription_schema}
\title{Get event subscription schema}
\usage{
get_event_subscription_schema(tier = "production")
}
\arguments{
\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) Schema for event subscription creation request entities
}
\description{
Get event subscription schema
}
\examples{
get_event_subscription_schema()

}
