% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_subscription_schema.R
\name{get_event_subscription_schema}
\alias{get_event_subscription_schema}
\title{Get event subscription schema}
\usage{
get_event_subscription_schema(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
("xml_document" "xml_node") EML metadata.
}
\description{
Get Event Subscription Schema operation, returns the XML schema for event subscription creation request entities.
}
\details{
GET : https://pasta.lternet.edu/package/event/eml/schema
}
\examples{

}
