% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_event_subscription_schema.R
\name{get_event_subscription_schema}
\alias{get_event_subscription_schema}
\title{Get event subscription schema}
\usage{
get_event_subscription_schema(env = "production")
}
\arguments{
\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Schema for event subscription creation request entities
}
\description{
Get event subscription schema
}
\examples{
# Get schema
schema <- get_event_subscription_schema()
schema

# Show schema structure
xml2::xml_structure(schema)

}
