% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journal_citation.R
\name{get_journal_citation}
\alias{get_journal_citation}
\title{Get journal citation}
\usage{
get_journal_citation(journalCitationId, env = "production")
}
\arguments{
\item{journalCitationId}{(numeric) Journal citation identifier}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(xml_document) Journal citation
}
\description{
Get journal citation
}
\examples{
\dontrun{

# Get citation
journalCitation <- get_journal_citation(381)
journalCitation
#> {xml_document}
#> <journalCitation>
#> [1] <journalCitationId>381</journalCitationId>
#> [2] <packageId>edi.845.1</packageId>
#> [3] <principalOwner>uid=csmith,o=EDI,dc=edirepository,dc=org</principa ...
#> [4] <dateCreated>2021-05-27T13:23:14.981</dateCreated>
#> [5] <articleDoi>https://doi.org/10.1016/j.scitotenv.2021.148033</artic ...
#> [6] <articleTitle>Bioturbation frequency alters methane emissions from ...
#> [7] <articleUrl>https://doi.org/10.1016/j.scitotenv.2021.148033</artic ...
#> [8] <journalTitle>Science of the Total Environment</journalTitle>
#> [9] <relationType>IsCitedBy</relationType>
}
}
\seealso{
Other Journal Citations: 
\code{\link{create_journal_citation}()},
\code{\link{delete_journal_citation}()},
\code{\link{list_data_package_citations}()},
\code{\link{list_principal_owner_citations}()}
}
\concept{Journal Citations}
