% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journal_citation.R
\name{get_journal_citation}
\alias{get_journal_citation}
\title{Get journal citation}
\usage{
get_journal_citation(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
("xml_document" "xml_node") EML metadata.
}
\description{
Get Journal Citation operation, returns an XML metadata document for the journal citation with the specified integer ID value.
}
\details{
GET : https://pasta.lternet.edu/package/citation/eml/{journalCitationId}
}
\examples{
# Using curl to access the journal citation with identifier value 15:

}
