% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journal_citation.R
\name{get_journal_citation}
\alias{get_journal_citation}
\title{Get journal citation}
\usage{
get_journal_citation(journalCitationId, tier = "production")
}
\arguments{
\item{journalCitationId}{(numeric) Journal citation identifier}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) Journal citation
}
\description{
Get journal citation
}
\examples{
# Get first journal citation of data package edi.845.1
journalCitationIds <- list_data_package_citations("edi.845.1")
journalCitationId <- xml2::xml_text(
  xml2::xml_find_first(citations, ".//journalCitationId"))
get_journal_citation(journalCitationId)

}
