% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journal_citation.R
\name{get_journal_citation}
\alias{get_journal_citation}
\title{Get journal citation}
\usage{
get_journal_citation(journalCitationId, as = "data.frame", env = "production")
}
\arguments{
\item{journalCitationId}{(numeric) Journal citation identifier}

\item{as}{(character) Format of the returned object. Can be: "data.frame" 
or "xml".}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(data.frame or xml_document) Journal citation
}
\description{
Get journal citation
}
\examples{
\dontrun{

# Get citation
journalCitation <- get_journal_citation(381)
}
}
\seealso{
Other Journal Citations: 
\code{\link{create_journal_citation}()},
\code{\link{delete_journal_citation}()},
\code{\link{list_data_package_citations}()},
\code{\link{list_principal_owner_citations}()}
}
\concept{Journal Citations}
