% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packageid_reads.R
\name{get_packageid_reads}
\alias{get_packageid_reads}
\title{Get package ID reads}
\usage{
get_packageid_reads(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(xml_document) Summary of all the successful reads (total reads and
non-robot reads) of \code{packageId}
}
\description{
Get package ID reads
}
\examples{
\dontrun{

# Get packageId reads
resourceReads <- get_packageid_reads("knb-lter-sgs.817.17")
resourceReads
#> {xml_document}
#> <resourceReads>
#> [1] <resource>\n  <resourceId>https://pasta.lternet.edu/package/metada ...
#> [2] <resource>\n  <resourceId>https://pasta.lternet.edu/package/eml/kn ...
#> [3] <resource>\n  <resourceId>https://pasta.lternet.edu/package/report ...
#> [4] <resource>\n  <resourceId>https://pasta.lternet.edu/package/data/e ...

# Show first
resource <- xml2::xml_find_first(resourceReads, "resource")
resource
#> {xml_node}
#> <resource>
#> [1] <resourceId>https://pasta.lternet.edu/package/metadata/eml/knb-lte ...
#> [2] <resourceType>metadata</resourceType>
#> [3] <scope>knb-lter-sgs</scope>
#> [4] <identifier>817</identifier>
#> [5] <revision>17</revision>
#> [6] <totalReads>535</totalReads>
#> [7] <nonRobotReads>518</nonRobotReads>
}
}
\seealso{
Other Audit Manager Services: 
\code{\link{get_audit_count}()},
\code{\link{get_audit_record}()},
\code{\link{get_audit_report}()},
\code{\link{get_docid_reads}()},
\code{\link{get_recent_uploads}()}
}
\concept{Audit Manager Services}
