% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packageid_reads.R
\name{get_packageid_reads}
\alias{get_packageid_reads}
\title{Get package ID reads}
\usage{
get_packageid_reads(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) Summary of all the successful reads (total reads and non-robot reads) of \code{packageId}
}
\description{
Get package ID reads
}
\examples{
get_packageid_reads("knb-lter-sgs.817.17")

}
