% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provenance_metadata.R
\name{get_provenance_metadata}
\alias{get_provenance_metadata}
\title{Get provenance metadata}
\usage{
get_provenance_metadata(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Provenance metadata
}
\description{
Get provenance metadata
}
\examples{
get_provenance_metadata("knb-lter-pal.309.1")

}
