% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_uploads.R
\name{get_recent_uploads}
\alias{get_recent_uploads}
\title{Get recent uploads}
\usage{
get_recent_uploads(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
("xml_document" "xml_node") EML metadata.
}
\description{
Gets a list of zero or more audit records of either recently inserted or recently updated data packages, as specified in the request.
}
\details{
GET : https://pasta.lternet.edu/audit/recent-uploads
}
\examples{

}
