% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_test_package}
\alias{get_test_package}
\title{Get the first data package in the staging environment for testing}
\usage{
get_test_package()
}
\value{
(character) Data package ID of the form "scope.identifier.revision".
}
\description{
Get the first data package in the staging environment for testing
}
