% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_authorized.R
\name{is_authorized}
\alias{is_authorized}
\title{Is authorized}
\usage{
is_authorized(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
("xml_document" "xml_node") XML containing the the 
    data package ID, data package title, and data package URL.
}
\description{
Is Authorized (to read resource) operation, determines whether the user as defined in the authentication token has permission to read the specified data package resource.
}
\details{
GET : https://pasta.lternet.edu/package/reservations/eml

GET : https://pasta.lternet.edu/package/authz?resourceId={resource identifier}
}
\examples{

}
