% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_authorized.R
\name{is_authorized}
\alias{is_authorized}
\title{Is authorized}
\usage{
is_authorized(resourceId, tier = "production")
}
\arguments{
\item{resourceId}{(character) Identifier of an EDI Repository resource}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(logical) Whether the user as defined in the authentication token has permission to read the specified data package resource
}
\description{
Is authorized
}
\note{
User authentication is required (see \code{login()})
}
\examples{

}
