% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_authorized.R
\name{is_authorized}
\alias{is_authorized}
\title{Is authorized to read}
\usage{
is_authorized(resourceId, tier = "production")
}
\arguments{
\item{resourceId}{(character) Identifier of an EDI Repository resource}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(logical) TRUE if the authenticated user has permission to read the specified resource
}
\description{
Is authorized to read
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{
resourceId <- "https://pasta.lternet.edu/package/report/eml/knb-lter-sbc/6006/3"
is_authorized(resourceId)
}

}
