% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_active_reservations.R
\name{list_active_reservations}
\alias{list_active_reservations}
\title{List active reservations}
\usage{
list_active_reservations(tier = "production")
}
\arguments{
\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) The set of data package identifiers that users have actively reserved. Note that data package identifiers that have been successfully uploaded are no longer considered active reservations and thus are not included in this list.
}
\description{
List active reservations
}
\examples{
list_active_reservations()

}
