% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_active_reservations.R
\name{list_active_reservations}
\alias{list_active_reservations}
\title{List active reservations}
\usage{
list_active_reservations(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
("xml_document" "xml_node") XML containing the the 
    data package ID, data package title, and data package URL.
}
\description{
List Active Reservations operation, lists the set of data package identifiers that users have actively reserved in PASTA. Note that data package identifiers that have been successfully uploaded into PASTA are no longer considered active reservations and thus are not included in this list.
}
\details{
GET : https://pasta.lternet.edu/package/reservations/eml
}
\examples{
# Using curl to list active reservations:

}
