% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_descendants.R
\name{list_data_descendants}
\alias{list_data_descendants}
\title{List data descendants}
\usage{
list_data_descendants(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(xml_document) Descendants of \code{packageId}
}
\description{
Data descendants are data packages that are known to be
derived, in whole or in part, from the specified source data package.
}
\examples{
\dontrun{

# List descendants
dataDescendants <- list_data_descendants("knb-lter-bnz.501.17")
dataDescendants
#> {xml_document}
#> <dataDescendants>
#> [1] <dataDescendant>\n  <packageId>edi.275.1</packageId>\n  <title>&a ...
#> [2] <dataDescendant>\n  <packageId>edi.275.2</packageId>\n  <title>&a ...
#> [3] <dataDescendant>\n  <packageId>edi.275.3</packageId>\n  <title>&a ...
#> [4] <dataDescendant>\n  <packageId>edi.275.4</packageId>\n  <title>&a ...
#> [5] <dataDescendant>\n  <packageId>edi.275.5</packageId>\n  <title>&a ...
#> [6] <dataDescendant>\n  <packageId>edi.275.6</packageId>\n  <title>Ei ...

# Show first
xml2::xml_find_first(dataDescendants, "dataDescendant")
#> {xml_node}
#> <dataDescendant>
#> [1] <packageId>edi.275.1</packageId>
#> [2] <title>&amp;#xa;      Eight Mile Lake Research Watershed&amp;#x2c ...
#> [3] <url>https://pasta.lternet.edu/package/metadata/eml/edi/275/1</url>
}
}
\seealso{
Other Listing: 
\code{\link{list_data_entities}()},
\code{\link{list_data_package_identifiers}()},
\code{\link{list_data_package_revisions}()},
\code{\link{list_data_package_scopes}()},
\code{\link{list_data_sources}()},
\code{\link{list_deleted_data_packages}()},
\code{\link{list_recent_changes}()},
\code{\link{list_recent_uploads}()},
\code{\link{list_service_methods}()},
\code{\link{list_user_data_packages}()}
}
\concept{Listing}
