% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_descendants.R
\name{list_data_descendants}
\alias{list_data_descendants}
\title{List Data Descendants}
\usage{
list_data_descendants(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
("xml_document" "xml_node") XML containing the the 
    data package ID, data package title, and data package URL.
}
\description{
List Data Descendants operation, specifying the scope, identifier, and revision values to match in the URI. Data descendants are data packages that are known to be derived, in whole or in part, from the specified source data package.
}
\details{
GET : https://pasta.lternet.edu/package/descendants/eml/{scope}/{identifier}/{revision}
}
\examples{
# Using curl to list data descendants of a data package:

}
