% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_descendants.R
\name{list_data_descendants}
\alias{list_data_descendants}
\title{List data descendants}
\usage{
list_data_descendants(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Descendants of \code{packageId}
}
\description{
Data descendants are data packages that are known to be derived, in whole or in part, from the specified source data package.
}
\examples{
list_data_descendants("knb-lter-bnz.501.17")

}
