% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_entities.R
\name{list_data_entities}
\alias{list_data_entities}
\title{List data entities}
\usage{
list_data_entities(scope, identifier, revision, environment = "production")
}
\arguments{
\item{scope}{(character) Data package scope}

\item{identifier}{(character) Data package identifier}

\item{revision}{(character) Data package revision}

\item{environment}{(character) PASTA environment to which this operation will be applied. Can be: 'production', 'staging', or 'development'}
}
\value{
(character) Data entity identifier(s) of the specified data package
}
\description{
List data entities
}
\details{
GET : https://pasta.lternet.edu/package/data/eml/{scope}/{identifier}/{revision}
}
\examples{
list_data_entities(scope = "edi", identifier = "193", revision = "5")

}
