% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_entities.R
\name{list_data_entities}
\alias{list_data_entities}
\title{List data entities}
\usage{
list_data_entities(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(character) Identifiers for all data entities in \code{packageId}
}
\description{
List data entities
}
\examples{
entityIds <- list_data_entities("knb-lter-and.2732.7")
entityIds

}
