% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_entities.R
\name{list_data_entities}
\alias{list_data_entities}
\title{List data entities}
\usage{
list_data_entities(package.id, environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(data frame) A data frame with columns:
    \itemize{
        \item{identifier: Data entity identifier}
    }
}
\description{
List Data Entities operation, specifying the scope, identifier, and 
    revision values to match in the URI.
}
