% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_citations.R
\name{list_data_package_citations}
\alias{list_data_package_citations}
\title{List data package citations}
\usage{
list_data_package_citations(packageId, env = "production", list_all = FALSE)
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}

\item{list_all}{(logical) Return all citations within a data package series?}
}
\value{
(xml_document) A list of journal citations
}
\description{
List data package citations
}
\examples{
\dontrun{

# List citations
journalCitations <- list_data_package_citations("edi.845.1")
journalCitations
#> {xml_document}
#> <journalCitations>
#> [1] <journalCitation>\n  <journalCitationId>381</journalCitationId>\n  ...

# Show first
xml2::xml_find_first(journalCitations, "journalCitation")
#> {xml_node}
#> <journalCitation>
#> [1] <journalCitationId>381</journalCitationId>
#> [2] <packageId>edi.845.1</packageId>
#> [3] <principalOwner>uid=csmith,o=EDI,dc=edirepository,dc=org</principa ...
#> [4] <dateCreated>2021-05-27T13:23:14.981</dateCreated>
#> [5] <articleDoi>https://doi.org/10.1016/j.scitotenv.2021.148033</artic ...
#> [6] <articleTitle>Bioturbation frequency alters methane emissions from ...
#> [7] <articleUrl>https://doi.org/10.1016/j.scitotenv.2021.148033</artic ...
#> [8] <journalTitle>Science of the Total Environment</journalTitle>
#> [9] <relationType>IsCitedBy</relationType>
}
}
\seealso{
Other Journal Citations: 
\code{\link{create_journal_citation}()},
\code{\link{delete_journal_citation}()},
\code{\link{get_journal_citation}()},
\code{\link{list_principal_owner_citations}()}
}
\concept{Journal Citations}
