% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_citations.R
\name{list_data_package_citations}
\alias{list_data_package_citations}
\title{List data package citations}
\usage{
list_data_package_citations(packageId, tier = "production", list_all = FALSE)
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}

\item{list_all}{(logical) Return all citations within a data package series?}
}
\value{
(xml_document) A list of journal citations
}
\description{
List data package citations
}
\examples{
list_data_package_citations("edi.845.1")

}
