% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_citations.R
\name{list_data_package_citations}
\alias{list_data_package_citations}
\title{List data package citations}
\usage{
list_data_package_citations(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) A list of journal citations
}
\description{
List data package citations
}
\examples{
list_data_package_citations("edi.845.1")

}
