% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_citations.R
\name{list_data_package_citations}
\alias{list_data_package_citations}
\title{List data package citations}
\usage{
list_data_package_citations(package.id, environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('XMLInternalDocument' 'XMLAbstractDocument') journalCitations metadata
}
\description{
List Data Package Citations operation, specifying the data package 
    scope, identifier, and revision values to match in the URI. 
    Returns a list of journal citations as an XML metadata document.
}
