% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_citations.R
\name{list_data_package_citations}
\alias{list_data_package_citations}
\title{List data package citations}
\usage{
list_data_package_citations(packageId, env = "production", list_all = FALSE)
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}

\item{list_all}{(logical) Return all citations within a data package series?}
}
\value{
(xml_document) A list of journal citations
}
\description{
List data package citations
}
\examples{
# List citations
journalCitations <- list_data_package_citations("edi.845.1")
journalCitations

# Show first
xml2::xml_find_first(journalCitations, "journalCitation")
}
\seealso{
Other Journal Citations: 
\code{\link{create_journal_citation}()},
\code{\link{delete_journal_citation}()},
\code{\link{get_journal_citation}()},
\code{\link{list_principal_owner_citations}()}
}
\concept{Journal Citations}
