% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_identifiers.R
\name{list_data_package_identifiers}
\alias{list_data_package_identifiers}
\title{List data package identifiers}
\usage{
list_data_package_identifiers(scope, environment = "production")
}
\arguments{
\item{scope}{(character) Data package scope (e.g. "edi", "knb-lter-bnz").}

\item{environment}{(character) PASTA environment to which this operation will be applied. Can be: "production", "staging", or "development".}
}
\value{
(character) Data package identifiers
}
\description{
List data package identifiers
}
\details{
GET : https://pasta.lternet.edu/package/eml/{scope}
}
\examples{
list_data_package_identifiers("knb-lter-ble")

}
