% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_identifiers.R
\name{list_data_package_identifiers}
\alias{list_data_package_identifiers}
\title{List data package identifiers}
\usage{
list_data_package_identifiers(scope, tier = "production")
}
\arguments{
\item{scope}{(character) Scope of data package (i.e. the first component of a \code{packageId})}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(numeric) Identifiers of data packages within a specified \code{scope}
}
\description{
List data package identifiers
}
\examples{
list_data_package_identifiers("knb-lter-ble")

}
