% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_identifiers.R
\name{list_data_package_identifiers}
\alias{list_data_package_identifiers}
\title{List data package identifiers}
\usage{
list_data_package_identifiers(scope, environment = 'production')
}
\arguments{
\item{scope}{(character) Scope (e.g. 'edi', 'knb-lter-bnz').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Data package identifiers
}
\description{
List Data Package Identifiers operation, specifying the scope value to 
    match in the URI.
}
