% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_identifiers.R
\name{list_data_package_identifiers}
\alias{list_data_package_identifiers}
\title{List data package identifiers}
\usage{
list_data_package_identifiers(scope, env = "production")
}
\arguments{
\item{scope}{(character) Scope of data package}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(numeric) Identifiers of data packages within a specified
\code{scope}
}
\description{
List data package identifiers
}
\examples{
# List identifiers
identifiers <- list_data_package_identifiers("knb-lter-ble")
identifiers
}
