% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_revisions.R
\name{list_data_package_revisions}
\alias{list_data_package_revisions}
\title{List data package revisions}
\usage{
list_data_package_revisions(
  scope,
  identifier,
  filter = NULL,
  environment = "production"
)
}
\arguments{
\item{scope}{(character) Data package scope (e.g. "edi", "knb-lter-bnz")}

\item{identifier}{(character) Data package identifier}

\item{filter}{(character) Filter returned revisions. Can be "newest" or "oldest"}

\item{environment}{(character) PASTA environment to which this operation will be applied. Can be: "production", "staging", or "development"}
}
\value{
(data.frame) Data package revision(s)
}
\description{
List data package revisions
}
\details{
GET : https://pasta.lternet.edu/package/eml/{scope}/{identifier}
}
\examples{
# All revisions
list_data_package_revisions("edi", "275")

# Newest revision
list_data_package_revisions("edi", "275", filter = "newest")

# Oldest revision
list_data_package_revisions("edi", "275", filter = "oldest")

}
