% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_revisions.R
\name{list_data_package_revisions}
\alias{list_data_package_revisions}
\title{List data package revisions}
\usage{
list_data_package_revisions(
  scope,
  identifier,
  filter = NULL,
  env = "production"
)
}
\arguments{
\item{scope}{(character) Scope of data package}

\item{identifier}{(numeric) Identifier of data package}

\item{filter}{(character) Filter results by "newest" or "oldest"}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(numeric) Revisions of a data package within a specified \code{scope} and \code{identifier}
}
\description{
List data package revisions
}
\examples{
# List revisions
revisions <- list_data_package_revisions("knb-lter-arc", 20131)
revisions

}
