% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_revisions.R
\name{list_data_package_revisions}
\alias{list_data_package_revisions}
\title{List data package revisions}
\usage{
list_data_package_revisions(
  scope,
  identifier,
  tier = "production",
  filter = NULL
)
}
\arguments{
\item{scope}{(character) Scope of data package (i.e. the first component of a \code{packageId})}

\item{identifier}{(numeric) Identifier of data package (i.e. the second component of a \code{packageId})}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}

\item{filter}{(character) Filter results by "newest" or "oldest"}
}
\value{
(numeric) Revisions of a data package within a specified \code{scope} and \code{identifier}
}
\description{
List data package revisions
}
\examples{
# All revisions
list_data_package_revisions("knb-lter-arc", 20131)

# Newest revision
list_data_package_revisions("knb-lter-arc", 20131, filter = "newest")

# Oldest revision
list_data_package_revisions("knb-lter-arc", 20131, filter = "oldest")

}
