% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_scopes.R
\name{list_data_package_scopes}
\alias{list_data_package_scopes}
\title{List data package scopes}
\usage{
list_data_package_scopes(tier = "production")
}
\arguments{
\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(numeric) Scopes within a specified \code{tier}
}
\description{
List data package scopes
}
\examples{
# All scopes in production
list_data_package_scopes()

# All scopes in staging
list_data_package_scopes("staging")

#' # All scopes in development
list_data_package_scopes("development")

}
