% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_package_scopes.R
\name{list_data_package_scopes}
\alias{list_data_package_scopes}
\title{List data package scopes}
\usage{
list_data_package_scopes(environment = "production")
}
\arguments{
\item{environment}{(character) PASTA environment to which this operation will be applied. Can be: "production", "staging", or "development"}
}
\value{
(character) All scope values extant in the data package registry.
}
\description{
List data package scopes
}
\details{
GET : https://pasta.lternet.edu/package/eml
}
\examples{
list_data_package_scopes()

}
