% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_sources.R
\name{list_data_sources}
\alias{list_data_sources}
\title{List Data Sources}
\usage{
list_data_sources(package.id, environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('xml_document' 'xml_node') For each data source, its 
    package identifier, title, and URL values are included (if applicable) 
    as documented in the metadata for the specified data package. Internal 
    data sources include a “packageId” value and a URL to the source 
    metadata. For data sources external to PASTA, the “packageId” element 
    will be empty and a URL value may or not be documented.
}
\description{
List Data Sources operation, specifying the scope, identifier, and revision values in the URI. For each data source, its package identifier, title, and URL values are included (if applicable) as documented in the metadata for the specified data package. Data sources can be either internal or external to PASTA. Internal data sources include a “packageId” value and a URL to the source metadata. For data sources external to PASTA, the “packageId” element will be empty and a URL value may or not be documented.
}
\details{
GET : https://pasta.lternet.edu/package/sources/eml/{scope}/{identifier}/{revision}
}
\examples{
# Using curl to list data sources of a data package:
}
