% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_sources.R
\name{list_data_sources}
\alias{list_data_sources}
\title{List data sources}
\usage{
list_data_sources(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and revision (e.g. 'edi.101.1').}

\item{environment}{(character) PASTA environment to which this operation will be applied. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) For each data source, its package identifier, title, and URL values are included (if applicable) as documented in the metadata for the specified data package.
}
\description{
Data sources can be either internal or external to PASTA. Internal data sources include a "packageId" value and a URL to the source metadata. For data sources external to PASTA, the "packageId" element will be empty and a URL value may or not be documented.
}
\details{
GET : https://pasta.lternet.edu/package/sources/eml/{scope}/{identifier}/{revision}
}
\examples{
list_data_sources("edi.275.4")

}
