% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_sources.R
\name{list_data_sources}
\alias{list_data_sources}
\title{List data sources}
\usage{
list_data_sources(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) Data sources to \code{packageId} including their data package identifier, title, and url
}
\description{
Data sources are data packages, or other online digital objects, that are known to be inputs to the specified derived data package.
}
\examples{
list_data_sources("edi.275.4")

}
