% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_deleted_data_packages.R
\name{list_deleted_data_packages}
\alias{list_deleted_data_packages}
\title{List deleted data packages}
\usage{
list_deleted_data_packages(tier = "production")
}
\arguments{
\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(character) All data packages (excluding revision values) that have been deleted from the data package registry.
}
\description{
List deleted data packages
}
\examples{
# All packages deleted from production
list_deleted_data_packages()

# All packages deleted from staging
list_deleted_data_packages("staging")

# All packages deleted from development
list_deleted_data_packages("development")

}
