% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_deleted_data_packages.R
\name{list_deleted_data_packages}
\alias{list_deleted_data_packages}
\title{List deleted data packages}
\usage{
list_deleted_data_packages(env = "production")
}
\arguments{
\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(character) All data packages (excluding revision values) that have
been deleted from the data package registry.
}
\description{
List deleted data packages
}
\examples{
# List deleted data packages
deleted <- list_deleted_data_packages()
head(deleted)
}
