% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_deleted_data_packages.R
\name{list_deleted_data_packages}
\alias{list_deleted_data_packages}
\title{List deleted data packages}
\usage{
list_deleted_data_packages(environment = "production")
}
\arguments{
\item{environment}{(character) PASTA environment to which this operation will be applied. Can be: "production", "staging", or "development"}
}
\value{
(data.frame) All document identifiers (excluding revision values) that have been deleted from the data package registry.
}
\description{
List deleted data packages
}
\details{
GET : https://pasta.lternet.edu/package/eml/deleted
}
\examples{
list_deleted_data_packages()

}
