% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_deleted_data_packages.R
\name{list_deleted_data_packages}
\alias{list_deleted_data_packages}
\title{List deleted data packages}
\usage{
list_deleted_data_packages(env = "production")
}
\arguments{
\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(character) All data packages (excluding revision values) that have
been deleted from the data package registry.
}
\description{
List deleted data packages
}
\examples{
\dontrun{

# List deleted data packages
deleted <- list_deleted_data_packages()
head(deleted)
#> [1] "edi.10"  "edi.222" "edi.419" "edi.511" "edi.857" "edi.878"
}
}
\seealso{
Other Listing: 
\code{\link{list_data_descendants}()},
\code{\link{list_data_entities}()},
\code{\link{list_data_package_identifiers}()},
\code{\link{list_data_package_revisions}()},
\code{\link{list_data_package_scopes}()},
\code{\link{list_data_sources}()},
\code{\link{list_recent_changes}()},
\code{\link{list_recent_uploads}()},
\code{\link{list_service_methods}()},
\code{\link{list_user_data_packages}()}
}
\concept{Listing}
