% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_principal_owner_citations.R
\name{list_principal_owner_citations}
\alias{list_principal_owner_citations}
\title{List principal owner citations}
\usage{
list_principal_owner_citations(dn, environment = "production")
}
\arguments{
\item{dn}{(character) Distinguished name (e.g. 
'uid=csmith,o=LTER,dc=ecoinformatics,dc=org')}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('xml_document' 'xml_node') journalCitations metadata
}
\description{
List Principal Owner Citations operation, returns journal citations metadata for all entries owned by the specified principal owner.
}
\details{
GET : https://pasta.lternet.edu/package/citations/eml/{principalOwner}
}
\examples{
# Using curl to access the list of journal citations owned by user “ucarroll”
}
