% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_principal_owner_citations.R
\name{list_principal_owner_citations}
\alias{list_principal_owner_citations}
\title{List principal owner citations}
\usage{
list_principal_owner_citations(principalOwner, tier = "production")
}
\arguments{
\item{principalOwner}{(character) Principal owner in the form returned by \code{construct_dn()}}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) Journal citations metadata for all entries owned by the specified principal owner
}
\description{
List principal owner citations
}
\examples{
principalOwner <- construct_dn("csmith")
list_principal_owner_citations(principalOwner)

}
