% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_recent_changes.R
\name{list_recent_changes}
\alias{list_recent_changes}
\title{List recent changes}
\usage{
list_recent_changes(
  fromDate = NULL,
  toDate = NULL,
  scope = NULL,
  environment = "production"
)
}
\arguments{
\item{fromDate}{(character) Start date (e.g. "2019-01-01T12:00:00")}

\item{toDate}{(character) End date (e.g. "2019-01-25T12:00:00")}

\item{scope}{(character) Data package scope}

\item{environment}{(character) PASTA environment to which this operation will be applied. Can be: "production", "staging", or "development"}
}
\value{
(xml_document) packageId, scope, identifier, revision, principal, doi, serviceMethod, and date of data package changes.
}
\description{
List all data package insert, update, and delete operations, optionally specifying the date and time to and/or from which the changes should be listed. An optional scope value can be specified to filter results for a particular data package scope.
}
\details{
GET : https://pasta.lternet.edu/package/changes/eml

If "fromDate" and "toDate" are omitted, lists the complete set of changes recorded in PASTA's  resource registry. If a "scope" value is omitted, results are returned for all data package scopes that exist in the resource registry. Multiple instances of the scope parameter are not supported. Inserts and updates are recorded in "dataPackageUpload" elements, while deletes are recorded in "dataPackageDelete" elements.
}
\examples{
# Changes occurring in January 2021 for the scope "knb-lter-hbr"
list_recent_changes(
  fromDate = "2021-01-01T00:00:00", 
  toDate = "2021-02-01T00:00:00", 
  scope = "knb-lter-hbr")
  
# Changes occurring in January 2021 for all scopes
list_recent_changes(
  fromDate = "2021-01-01T00:00:00", 
  toDate = "2021-02-01T00:00:00")

}
