% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_recent_changes.R
\name{list_recent_changes}
\alias{list_recent_changes}
\title{List recent changes}
\usage{
list_recent_changes(
  fromDate = NULL,
  toDate = NULL,
  scope = NULL,
  env = "production"
)
}
\arguments{
\item{fromDate}{(character) Start date in the format "YYYY-MM-DDThh:mm:ss"}

\item{toDate}{(character) End date in the format "YYYY-MM-DDThh:mm:ss"}

\item{scope}{(character) Scope of data package}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Recent changes and their corresponding packageId, scope, identifier, revision, principal, doi, serviceMethod, and date.
}
\description{
List all data package insert, update, and delete operations, optionally specifying the date and time to and/or from which the changes should be listed. An optional scope value can be specified to filter results for a particular data package scope.
}
\examples{
# Changes occurring in January 2021 within the scope "knb-lter-hbr"
list_recent_changes(
  fromDate = "2021-01-01T00:00:00", 
  toDate = "2021-02-01T00:00:00", 
  scope = "knb-lter-hbr")
  
# Changes occurring in the first 3 days of 2021 for all scopes
list_recent_changes(
  fromDate = "2021-01-01T00:00:00", 
  toDate = "2021-01-03T00:00:00")

}
