% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_recent_uploads.R
\name{list_recent_uploads}
\alias{list_recent_uploads}
\title{List recent uploads}
\usage{
list_recent_uploads(type, limit = 5, tier = "production")
}
\arguments{
\item{type}{(character) Upload type, which can be: "insert" or "update"}

\item{limit}{(numeric) Limit on return values}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) Data package uploads and their packageId, scope, identifier, revision, principal, doi, serviceMethod, and date.
}
\description{
List recent uploads
}
\examples{
# Get the 5 newest data packages
list_recent_uploads("insert", 3)

# Get the 5 newest revisions
list_recent_uploads("update", 3)

}
