% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_recent_uploads.R
\name{list_recent_uploads}
\alias{list_recent_uploads}
\title{List recent uploads}
\usage{
list_recent_uploads(type, limit = 5, env = "production")
}
\arguments{
\item{type}{(character) Upload type. Can be: "insert" or "update".}

\item{limit}{(numeric) Maximum number of results to return}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Data package uploads
}
\description{
List recent uploads
}
\examples{
# Get the 5 newest data packages
list_recent_uploads("insert", 3)

# Get the 5 newest revisions
list_recent_uploads("update", 3)

}
