% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_recent_uploads.R
\name{list_recent_uploads}
\alias{list_recent_uploads}
\title{List recent uploads}
\usage{
list_recent_uploads(type, limit = 5, environment = "production")
}
\arguments{
\item{type}{(character) Upload type. Can be: "insert" or "update".}

\item{limit}{(integer) Maximum limit.}

\item{environment}{(character) PASTA environment to which this operation will be applied. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) packageId, scope, identifier, revision, principal, doi, serviceMethod, and date of data package uploads.
}
\description{
List recent uploads
}
\details{
GET : https://pasta.lternet.edu/package/uploads/eml
}
\examples{
list_recent_uploads("update", 10)

}
