% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_reservation_identifiers.R
\name{list_reservation_identifiers}
\alias{list_reservation_identifiers}
\title{List reservation identifiers}
\usage{
list_reservation_identifiers(type, limit = 5, environment = "production")
}
\arguments{
\item{type}{(character) Upload type. Can be: "insert" or "update".}

\item{limit}{(integer) Maximum limit.}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('xml_document' 'xml_node') dataPackageUploads 
    metadata.
}
\description{
List Reservation Identifiers operation, lists the set of numeric identifiers for the specified scope that end users have actively reserved for future upload to PASTA. The numeric identifiers are listed one per line.
}
\details{
GET : https://pasta.lternet.edu/package/reservations/eml/{scope}
}
\examples{
# Using curl to list reservation identifiers for a specified scope:

}
