% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_reservation_identifiers.R
\name{list_reservation_identifiers}
\alias{list_reservation_identifiers}
\title{List reservation identifiers}
\usage{
list_reservation_identifiers(scope, tier = "production")
}
\arguments{
\item{scope}{(character) Scope of data package}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
(numeric) The set of identifiers for the specified scope that end users have actively reserved for future upload
}
\description{
List reservation identifiers
}
\examples{
list_reservation_identifiers("edi")

}
