% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service_methods.R
\name{list_service_methods}
\alias{list_service_methods}
\title{List service methods}
\usage{
list_service_methods(environment = "production")
}
\arguments{
\item{environment}{(character) PASTA environment to which this operation will be applied. Can be: "production", "staging", or "development"}
}
\value{
(data.frame) A simple list of web service methods supported by the Data Package Manager web service.
}
\description{
List service methods
}
\details{
GET : https://pasta.lternet.edu/package/service-methods
}
\examples{
list_service_methods()

}
