% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service_methods.R
\name{list_service_methods}
\alias{list_service_methods}
\title{List service methods}
\usage{
list_service_methods(env = "production")
}
\arguments{
\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(character) A simple list of web service methods supported by the
Data Package Manager web service
}
\description{
List service methods
}
\examples{
# All service methods
services <- list_service_methods()
services
}
\seealso{
Other Listing: 
\code{\link{list_data_descendants}()},
\code{\link{list_data_entities}()},
\code{\link{list_data_package_identifiers}()},
\code{\link{list_data_package_revisions}()},
\code{\link{list_data_package_scopes}()},
\code{\link{list_data_sources}()},
\code{\link{list_deleted_data_packages}()},
\code{\link{list_recent_changes}()},
\code{\link{list_recent_uploads}()},
\code{\link{list_user_data_packages}()}
}
\concept{Listing}
