% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service_methods.R
\name{list_service_methods}
\alias{list_service_methods}
\title{List service methods}
\usage{
list_service_methods(tier = "production")
}
\arguments{
\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(character) A simple list of web service methods supported by the Data Package Manager web service.
}
\description{
List service methods
}
\examples{
# All service methods in production
list_service_methods()

# All service methods in staging
list_service_methods("staging")

#' # All service methods in development
list_service_methods("development")

}
