% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_service_methods.R
\name{list_service_methods}
\alias{list_service_methods}
\title{List service methods}
\usage{
list_service_methods(environment = 'production')
}
\arguments{
\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) All scope values extant in the data package registry.
}
\description{
List Service Methods operation, returning a simple list of web service 
    methods supported by the Data Package Manager web service.
}
