% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_user_data_packages.R
\name{list_user_data_packages}
\alias{list_user_data_packages}
\title{List user data packages}
\usage{
list_user_data_packages(dn, environment = "production")
}
\arguments{
\item{dn}{(character) Distinguished name (e.g. "uid=dbjourneynorth,o=EDI,dc=edirepository,dc=org"). Use \code{get_distinguished_name()} to create input to \code{dn}.}

\item{environment}{(character) PASTA environment to which this operation will be applied. Can be: "production", "staging", or "development"}
}
\value{
(character) Data packages uploaded by a distinguished name.
}
\description{
List all data packages (including their revision values) uploaded to the repository by a particular user, specified by a distinguished name. Data packages that were uploaded by the specified user but have since been deleted are excluded from the list.
}
\details{
GET : https://pasta.lternet.edu/package/user/{dn}
}
\examples{
# List all (undeleted) data packages uploaded by user "dbjourneynorth" under the organizational unit "EDI"
list_user_data_packages(get_distinguished_name("dbjourneynorth", "EDI"))

}
