% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_user_data_packages.R
\name{list_user_data_packages}
\alias{list_user_data_packages}
\title{List user data packages}
\usage{
list_user_data_packages(dn, environment = 'production')
}
\arguments{
\item{dn}{(character) Distinguished name (e.g. 'uid=csmith,o=LTER,dc=ecoinformatics,dc=org')}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Data packages uploaded by a distinguished name.
}
\description{
List all data packages (including their revision values) uploaded to the 
    repository by a particular user, specified by a distinguished name. Data 
    packages that were uploaded by the specified user but have since been 
    deleted are excluded from the list.
}
