% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_user_data_packages.R
\name{list_user_data_packages}
\alias{list_user_data_packages}
\title{List user data packages}
\usage{
list_user_data_packages(dn, tier = "production")
}
\arguments{
\item{dn}{(character) Distinguished name of user. Use \code{create_dn()} to create input to this argument.}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(character) Data package identifiers belonging to a \code{dn}
}
\description{
List all data packages (including their revision values) uploaded to the repository by a particular user, specified by a distinguished name. Data packages that were uploaded by the specified user but have since been deleted are excluded from the list.
}
\examples{
# List all data packages uploaded by user "dbjourneynorth" who is apart of the "EDI" organizational unit
dn <- create_dn("dbjourneynorth", "EDI")
list_user_data_packages(dn)

}
