% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_working_on.R
\name{list_working_on}
\alias{list_working_on}
\title{List working on}
\usage{
list_working_on(environment = 'production')
}
\arguments{
\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('XMLInternalDocument' 'XMLAbstractDocument') workingOn metadata
}
\description{
List Working On operation, lists the set of data packages that PASTA is 
    currently working on inserting or updating. (Note that data packages 
    currently being evaluated by PASTA are not included in the list.)
}
