% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_working_on.R
\name{list_working_on}
\alias{list_working_on}
\title{List working on}
\usage{
list_working_on(as = "data.frame", env = "production")
}
\arguments{
\item{as}{(character) Format of the returned object. Can be: "data.frame" 
or "xml".}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(data.frame or xml_document) The set of data packages the EDI 
repository is currently working on inserting or updating. Note that data 
packages currently being evaluated by the EDI repository are not included in 
the list.
}
\description{
List working on
}
\examples{
\dontrun{

list_working_on()
}
}
\concept{System Monitoring}
