% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to the EDI repository}
\usage{
login(userId = NULL, userPass = NULL, config = NULL)
}
\arguments{
\item{userId}{(character) User identifier of an EDI data repository account}

\item{userPass}{(character) Password of \code{userId}}

\item{config}{(character) Path to config.txt, which contains \code{userId} and \code{userPass} (see details below)}
}
\value{
(character) A temporary (~10 hour) authentication token written to the system variable "EDI_TOKEN".
}
\description{
Login to the EDI repository
}
\details{
If \code{userId}, \code{userPass}, and \code{config} are NULL, the console will prompt for credentials.

\code{config}: Supplying credentials in a file named config.txt facilitates authentication within automated/unassisted processes. Contents of this file should be new line separated and have the form "<argument> = <value>" (e.g. userId = myname).
}
\note{
Only works when authenticating with EDI credentials. Does not work for ORCiD, GitHub, or Google credentials.
}
\examples{
\dontrun{

# Login with arguments
login(userId = "myname", userPass = "mysecret")

# Login with config.txt
login(config = "./data/config.txt")

# Login at console
login()
}

}
