% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\title{Login to the EDI repository}
\usage{
login(userId = NULL, userPass = NULL, config = NULL)
}
\arguments{
\item{userId}{(character) User identifier of an EDI data repository account}

\item{userPass}{(character) Password of \code{userId}}

\item{config}{(character) Path to config.txt, which contains \code{userId} and \code{userPass} (see details below)}
}
\value{
(file) A temporary (~10 hour) authentication token written to edi_token.txt within the session \code{tempdir()}.
}
\description{
Login to the EDI repository
}
\details{
If \code{userId}, \code{userPass}, and \code{config} are NULL, the console will prompt for credentials.

\code{config}: Supplying credentials in a file named config.txt facilitates authentication within automated/unassisted processes. Contents of this file should be new line separated and have the form "<argument> = <value>" (e.g. userId = myname).

Security: Input user name and password are only held for the duration of this function call and are disposed of by \code{on.exit()}. The temporary encrypted token is written to file at the path defined by \code{tempdir()} and is removed at close of the users R session.
}
\note{
Only works when authenticating with EDI credentials. Does not work for ORCiD, GitHub, or Google credentials.
}
\examples{
\dontrun{
# Login with arguments
login("myname", "mysecret")

# Login with config.txt
login("/Users/me/Documents/config.txt")

# Login at console
login()
}

}
