% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logout}
\alias{logout}
\title{Logout of the EDI repository}
\usage{
logout()
}
\description{
Logout of the EDI repository
}
\details{
Removes the temporary authentication token from the path defined by \code{tempdir()}.
}
\examples{
\dontrun{
logout()
}

}
