% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logout}
\alias{logout}
\title{Logout of the EDI repository}
\usage{
logout()
}
\value{
(NULL) No return value.
}
\description{
Logout of the EDI repository
}
\details{
Removes the temporary authentication token system variable
"EDI_TOKEN".
}
\examples{
\dontrun{
logout()
}

}
\seealso{
Other Authentication: 
\code{\link{login}()}
}
\concept{Authentication}
