% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reservation.R, R/pkg_reserve_id.R
\name{poll_pkg_reserve_id}
\alias{poll_pkg_reserve_id}
\title{Polling loop for reserving package IDs}
\usage{
poll_pkg_reserve_id(r)

poll_pkg_reserve_id(r)
}
\arguments{
\item{r}{(response) Response generated by `httr::POST`.}

\item{r}{(response) Response generated by `httr::POST`.}
}
\value{
(character) Package identifier.

(character) Package identifier.
}
\description{
Polling loop for reserving package IDs.

Polling loop for reserving package IDs.
}
