% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_event_subscriptions.R
\name{query_event_subscriptions}
\alias{query_event_subscriptions}
\title{Query event subscriptions}
\usage{
query_event_subscriptions(scope, environment, user.id, user.pass, affiliation)
}
\arguments{
\item{scope}{(character) Scope of identifier to be reserved (e.g. edi, knb-lter-ntl).}

\item{environment}{(character) Data repository environment in which to reserve the
identifier. Can be: 'development', 'staging', 'production'.}

\item{user.id}{(character) Identification of user reserving the identifier.}

\item{user.pass}{(character) Password corresponding with the user.id argument supplied
above.}

\item{affiliation}{(character) Affiliation corresponding with the user.id argument supplied
above. Can be: 'LTER' or 'EDI'.}
}
\value{
(character) Package identifier.
}
\description{
Query Event Subscriptions operation, returns a list of the subscriptions whose attributes match those specified in the query string. If a query string is omitted, all subscriptions in the subscription database will be returned for which the requesting user is authorized to read. If query parameters are included, they are used to filter that set of subscriptions based on their attributes.
}
\details{
GET : https://pasta.lternet.edu/package/event/eml
}
\examples{

}
