% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity.R
\name{read_data_entity}
\alias{read_data_entity}
\title{Read data entity}
\usage{
read_data_entity(package.id, identifier, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{identifier}{(character) Data entity identifier (e.g. 
5c224a0e74547b14006272064dc869b1)}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Data entity checksum
}
\description{
Read Data Entity operation, specifying the scope, identifier, revision, and entity identifier of the data entity to be read in the URI. Revision may be specified as “newest” or “oldest” to retrieve data from the newest or oldest revision, respectively.
}
\details{
GET : https://pasta.lternet.edu/package/data/eml/{scope}/{identifier}/{revision}/{entityId}
}
\examples{

}
