% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity.R
\name{read_data_entity}
\alias{read_data_entity}
\title{Read data entity}
\usage{
read_data_entity(packageId, entityId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{entityId}{(character) Data entity identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(raw) Raw bytes (i.e. application/octet-stream) to be parsed by a reader function appropriate for the data type
}
\description{
Read data entity
}
\examples{
# Get data entity
entityId <- list_data_entities(packageId = "edi.993.1")
entityId

# Read raw bytes
raw <- read_data_entity(packageId = "edi.993.1", entityId = entityId)
head(raw)

# Parse with .csv reader
data <- readr::read_csv(file = raw)
data

}
