% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity.R
\name{read_data_entity}
\alias{read_data_entity}
\title{Read data entity}
\usage{
read_data_entity(packageId, entityId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{entityId}{(character) Data entity identifier}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(raw) Raw bytes (i.e. application/octet-stream) to be parsed by a reader function appropriate for the data type
}
\description{
Read data entity
}
\examples{
# Get raw bytes of a .csv and use parser of choice
packageId <- "edi.993.1"
entityId <- list_data_entities(packageId)
raw <- read_data_entity(packageId, entityId)
data <- readr::read_csv(raw)

}
