% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity_checksum.R
\name{read_data_entity_checksum}
\alias{read_data_entity_checksum}
\title{Read data entity checksum}
\usage{
read_data_entity_checksum(packageId, entityId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{entityId}{(character) Data entity identifier}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(character) A 40-character SHA-1 checksum value of \code{entityId}
in \code{packageId}
}
\description{
Read data entity checksum
}
\examples{
# List data entities
entityIds <- list_data_entities(packageId = "knb-lter-ble.1.7")
entityIds

# Read checksum
checksum <- read_data_entity_checksum(
  packageId = "knb-lter-ble.1.7",
  entityId = entityIds[1]
)
checksum
}
