% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity_checksum.R
\name{read_data_entity_checksum}
\alias{read_data_entity_checksum}
\title{Read data entity checksum}
\usage{
read_data_entity_checksum(package.id, entityId, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and revision (e.g. "edi.101.1").}

\item{entityId}{(character) Data entity identifier (e.g. 5c224a0e74547b14006272064dc869b1)}

\item{environment}{(character) PASTA environment to which this operation will be applied. Can be: "production", "staging", or "development"}
}
\value{
(character) Data entity checksum value
}
\description{
Read data entity checksum
}
\details{
GET : https://pasta.lternet.edu/package/data/checksum/eml/{scope}/{identifier}/{revision}/{entityId}
}
\examples{


}
