% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity_checksum.R
\name{read_data_entity_checksum}
\alias{read_data_entity_checksum}
\title{Read data entity checksum}
\usage{
read_data_entity_checksum(packageId, entityId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{entityId}{(character) Data entity identifier}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(character) Checksum value of \code{entityId} in \code{packageId}
}
\description{
Read data entity checksum
}
\examples{
list_data_entities("knb-lter-ble.1.7")

}
