% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity_name.R
\name{read_data_entity_name}
\alias{read_data_entity_name}
\title{Read data entity name}
\usage{
read_data_entity_name(package.id, identifier, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and revision (e.g. "edi.101.1")}

\item{environment}{(character) PASTA environment to which this operation will be applied. Can be: "production", "staging", or "development".}

\item{entityId}{(character) Data entity identifier (e.g. "5c224a0e74547b14006272064dc869b1")}
}
\value{
(character) Data entity name
}
\description{
Read data entity name
}
\details{
GET : https://pasta.lternet.edu/package/name/eml/{scope}/{identifier}/{revision}/{entityId}
}
\examples{

}
