% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity_name.R
\name{read_data_entity_name}
\alias{read_data_entity_name}
\title{Read data entity name}
\usage{
read_data_entity_name(packageId, entityId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{entityId}{(character) Data entity identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(character) Name of \code{entityId} in \code{packageId}
}
\description{
Read data entity name
}
\examples{
# Get name of first data entity in data package "knb-lter-cap.691.2"
packageId <- "knb-lter-cap.691.2"
entityIds <- list_data_entities(packageId)
read_data_entity_name(packageId, entityIds[1])

}
