% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity_names.R
\name{read_data_entity_names}
\alias{read_data_entity_names}
\title{Read data entity names}
\usage{
read_data_entity_names(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(data.frame) Names and identifiers of all data entities in \code{packageId}
}
\description{
Read data entity names
}
\examples{
read_data_entity_names("knb-lter-cap.691.2")

}
