% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity_names.R
\name{read_data_entity_names}
\alias{read_data_entity_names}
\title{Read data entity names}
\usage{
read_data_entity_names(package.id, identifier, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{identifier}{(character) Data entity identifier (e.g. 
5c224a0e74547b14006272064dc869b1)}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Data entity names
}
\description{
Read Data Entity Names operation, specifying the scope, identifier, and revision of the data package whose data entity names are to be read in the URI, returning a newline-separated list of entity identifiers and name values. Each line in the list contains an entity identifier and its corresponding name value, separated by a comma. Only data entities that the user is authorized to read are included in the list.
}
\details{
GET : https://pasta.lternet.edu/package/name/eml/{scope}/{identifier}/{revision}
}
\examples{

}
