% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity_resource_metadata.R
\name{read_data_entity_resource_metadata}
\alias{read_data_entity_resource_metadata}
\title{Read data entity resource metadata}
\usage{
read_data_entity_resource_metadata(
  package.id,
  entity.id,
  environment = "production"
)
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{entity.id}{(character) Data entity identifier (e.g. 
"2353ac38985edd6aff140e4c65cb32de")}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('xml_document' 'xml_node') An XML string representing 
    the resource metadata for the data entity.
}
\description{
Read Data Entity Resource Metadata operation, specifying the scope, identifier, revision, and entity identifier of the data entity object whose resource metadata is to be read in the URI, returning an XML string representing the resource metadata for the data entity.
}
\details{
GET : https://pasta.lternet.edu/package/data/rmd/eml/{scope}/{identifier}/{revision}/{entityId}
}
\examples{

}
