% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity_size.R
\name{read_data_entity_size}
\alias{read_data_entity_size}
\title{Read data entity size}
\usage{
read_data_entity_size(packageId, entityId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{entityId}{(character) Data entity identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(numeric) Size, in bytes, of \code{entityId} in \code{packageId}
}
\description{
Read data entity size
}
\examples{
# List data entities
entityIds <- list_data_entities(packageId = "knb-lter-cdr.711.1")
entityIds

# Read size
size <- read_data_entity_size(
  packageId = "knb-lter-cdr.711.1", 
  entityId = entityIds)
size

}
