% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity_size.R
\name{read_data_entity_size}
\alias{read_data_entity_size}
\title{Read data entity size}
\usage{
read_data_entity_size(package.id, identifier, 
    environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{identifier}{(character) Data entity identifier (e.g. 
5c224a0e74547b14006272064dc869b1)}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Data entity size (in bytes)
}
\description{
Read Data Entity Size operation, specifying the scope, identifier, and 
    revision of the data entity object whose size is to be read in the URI, 
    returning the size value (in bytes).
}
