% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity_sizes.R
\name{read_data_entity_sizes}
\alias{read_data_entity_sizes}
\title{Read data entity sizes}
\usage{
read_data_entity_sizes(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
(data.frame) Size (in bytes) and identifiers of data entities in \code{packageId}
}
\description{
Read data entity sizes
}
\examples{
read_data_entity_sizes("knb-lter-cdr.711.1")

}
