% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_entity_sizes.R
\name{read_data_entity_sizes}
\alias{read_data_entity_sizes}
\title{Read data entity sizes}
\usage{
read_data_entity_sizes(package.id, identifier, 
    environment = 'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Data entity sizes (in bytes)
}
\description{
Read Data Entity Sizes operation, specifying the scope, identifier, 
    and revision of the data package whose data entity sizes are to be read 
    in the URI, returning a newline-separated list of entity identifiers 
    and size values (in bytes). Only data entities that the user is 
    authorized to read are included in the list.
}
