% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package.R
\name{read_data_package}
\alias{read_data_package}
\title{Read data package}
\usage{
read_data_package(
  package.id,
  filter = NULL,
  ore = FALSE,
  environment = "production"
)
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and revision (e.g. "edi.101.1")}

\item{filter}{(character) Filter returned revisions. Can be "newest" or "oldest"}

\item{ore}{(logical) Return an OAI-ORE compliant resource map in RDF-XML format}

\item{environment}{(character) PASTA environment to which this operation will be applied. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) A resource map with reference URLs to each of the metadata, data, and quality report resources that comprise the data package.
}
\description{
Read data package
}
\details{
GET : https://pasta.lternet.edu/package/eml/{scope}/{identifier}/{revision}
}
\examples{
# Read a data package resource map

# Read a data package resource map in OAI-ORE compliant format

}
