% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package.R
\name{read_data_package}
\alias{read_data_package}
\title{Read data package}
\usage{
read_data_package(packageId, tier = "production", ore = FALSE)
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}

\item{ore}{(logical) Return an OAI-ORE compliant resource map in RDF-XML format}
}
\value{
(character or xml_document) A resource map with reference URLs to each of the metadata, data, and quality report resources that comprise the \code{packageId}.
}
\description{
Read data package
}
\examples{
# Get resource map
read_data_package("knb-lter-cwt.5026.13")

# Get resource map in ORE format
read_data_package("knb-lter-cwt.5026.13", ore = TRUE)

}
