% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package.R
\name{read_data_package}
\alias{read_data_package}
\title{Read data package}
\usage{
read_data_package(packageId, ore = FALSE, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{ore}{(logical) Return an OAI-ORE compliant resource map in RDF-XML
format}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(character or xml_document) A resource map with reference URLs to
each of the metadata, data, and quality report resources that comprise the
\code{packageId}.
}
\description{
Read data package
}
\examples{
# Get resource map
resourceMap <- read_data_package(packageId = "knb-lter-cwt.5026.13")
resourceMap

# Get resource map in ORE format
resourceMap <- read_data_package(
  packageId = "knb-lter-cwt.5026.13",
  ore = TRUE
)
resourceMap
}
