% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_acl.R
\name{read_data_package_acl}
\alias{read_data_package_acl}
\title{Read data package ACL}
\usage{
read_data_package_acl(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) The canonical Digital Object Identifier.
}
\description{
Read Data Package ACL operation, specifying the scope, identifier, and revision of the data package whose Access Control List (ACL) is to be read in the URI, returning an XML string representing the ACL for the data package. Please note: only a very limited set of users are authorized to use this service method.
}
\details{
GET : https://pasta.lternet.edu/package/acl/eml/{scope}/{identifier}/{revision}
}
\examples{

}
