% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_archive.R
\name{read_data_package_archive}
\alias{read_data_package_archive}
\title{Read data package archive}
\usage{
read_data_package_archive(packageId, transaction, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{transaction}{(character) Transaction identifier}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(.zip) The data package archive of \code{packageId}
}
\description{
Read data package archive
}
\examples{

}
