% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_archive.R
\name{read_data_package_archive}
\alias{read_data_package_archive}
\title{Read data package archive}
\usage{
read_data_package_archive(packageId, transaction, path, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{transaction}{(character) Transaction identifier}

\item{path}{(character) Path of directory in which the result will be written}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(.zip file) The data package archive of \code{packageId} requested by \code{transaction}
}
\description{
Read data package archive
}
\examples{
packageId <- "knb-lter-vcr.340.1"
transaction <- create_data_package_archive(packageId)
path <- "/Users/me/Documents"
read_data_package_archive(packageId, transaction, path)

}
