% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_archive.R
\name{read_data_package_archive}
\alias{read_data_package_archive}
\title{Read data package archive}
\usage{
read_data_package_archive(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Reference URLs.
}
\description{
Read Data Package Archive operation, specifying the transaction identifier of the data package archive to be read in the URI, returning the data package archive as a binary object in the ZIP file format.
}
\details{
GET : https://pasta.lternet.edu/package/archive/eml/{scope}/{identifier}/{revision}/{transaction}
}
\examples{
# Using curl to read a data package archive and redirect the output to a file:

}
