% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_doi.R
\name{read_data_package_doi}
\alias{read_data_package_doi}
\title{Read data package Digital Object Identifier}
\usage{
read_data_package_doi(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(character) The canonical Digital Object Identifier for
\code{packageId}
}
\description{
Read data package Digital Object Identifier
}
\examples{
# Read package DOI
doi <- read_data_package_doi(packageId = "knb-lter-jrn.210548103.15")
doi
}
