% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_doi.R
\name{read_data_package_doi}
\alias{read_data_package_doi}
\title{Read data package Digital Object Identifier}
\usage{
read_data_package_doi(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
(character) The canonical Digital Object Identifier for \code{packageId}
}
\description{
Read data package Digital Object Identifier
}
\examples{
read_data_package_doi("knb-lter-jrn.210548103.15")

}
