% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_error.R
\name{read_data_package_error}
\alias{read_data_package_error}
\title{Read data package error}
\usage{
read_data_package_error(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Reference URLs.
}
\description{
Read Data Package Error operation, specifying the scope, identifier, revision, and transaction identifier of the data package error to be read in the URI, returning the error message as plain text.
}
\details{
GET : https://pasta.lternet.edu/package/error/eml/{transaction}
}
\examples{

}
