% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_error.R
\name{read_data_package_error}
\alias{read_data_package_error}
\title{Read data package error}
\usage{
read_data_package_error(packageId, transaction, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{transaction}{(character) Transaction identifier returned for each data package evaluate, upload, and delete operation}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(character) Data package error
}
\description{
Read data package error
}
\examples{

}
