% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_error.R
\name{read_data_package_error}
\alias{read_data_package_error}
\title{Read data package error}
\usage{
read_data_package_error(transaction, tier = "production")
}
\arguments{
\item{transaction}{(character) Transaction identifier}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(character) Result of operation
}
\description{
Read data package error
}
\examples{

}
