% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_error.R
\name{read_data_package_error}
\alias{read_data_package_error}
\title{Read data package error}
\usage{
read_data_package_error(transaction, env = "production")
}
\arguments{
\item{transaction}{(character) Transaction identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
An error is returned if an error occurred while processing the request, otherwise \code{NULL} is returned if no error was encountered or if processing is still underway.
}
\description{
Read data package error
}
\note{
User authentication is required (see \code{login()})
}
\examples{
\dontrun{

# Create zip archive
transaction <- create_data_package_archive(packageId = "knb-lter-sev.31999.1")
transaction
#> [1] "archive_knb-lter-sev.31999.1_16396683904724129"

# Check creation status
read_data_package_error(transaction)

# Download zip archive
read_data_package_archive(
  packageId = "knb-lter-sev.31999.1", 
  transaction = transaction, 
  path = "./data")
#> |=============================================================| 100\%
dir("./data")
#> [1] "knb-lter-sev.31999.1.zip"
}

}
