% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_from_doi.R
\name{read_data_package_from_doi}
\alias{read_data_package_from_doi}
\title{Read data package from DOI}
\usage{
read_data_package_from_doi(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Reference URLs.
}
\description{
Read Data Package From DOI operation, specifying the DOI of the data package to be read in the URI, returning a resource map with reference URLs to each of the metadata, data, and quality report resources that comprise the data package. The DOI is specified in the “shoulder”, “pasta”, and “md5” path segments of the URI (see example below). When the “?ore” query parameter is appended to the request URL, an OAI-ORE compliant resource map in RDF-XML format is returned.
}
\details{
GET : https://pasta.lternet.edu/package/doi/{shoulder}/{pasta}/{md5}
}
\examples{
# Using curl to read a data package resource map by specifying the DOI as three path segments in the URL:

}
