% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_report.R
\name{read_data_package_report}
\alias{read_data_package_report}
\title{Read data package report}
\usage{
read_data_package_report(
  packageId,
  tier = "production",
  html = FALSE,
  strip_ns = TRUE
)
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}

\item{html}{(logical) Return result in HTML format?}

\item{strip_ns}{(logical) Strip result namespace?}
}
\value{
(xml_document or html) Data package report
}
\description{
Read data package report
}
\examples{
# Result in XML format
read_data_package_report("knb-lter-knz.260.4")

# Result in HTML format
read_data_package_report("knb-lter-knz.260.4", html = TRUE)

}
