% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_report.R
\name{read_data_package_report}
\alias{read_data_package_report}
\title{Read data package report}
\usage{
read_data_package_report(packageId, html = FALSE, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{html}{(logical) Return result in HTML format?}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Data package report
}
\description{
Read data package report
}
\examples{
# Read report
qualityReport <- read_data_package_report(packageId = "knb-lter-knz.260.4")
qualityReport

}
