% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_report.R
\name{read_data_package_report}
\alias{read_data_package_report}
\title{Read data package report}
\usage{
read_data_package_report(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('xml_document' 'xml_node') Data package report.
}
\description{
Read Data Package Report operation, specifying the scope, identifier, and revision of the data package quality report document to be read in the URI. If an HTTP Accept header with value ‘text/html’ is included in the request, returns an HTML representation of the report. The default representation is XML.
}
\details{
GET : https://pasta.lternet.edu/package/report/eml/{scope}/{identifier}/{revision}
}
\examples{

}
