% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_report_acl.R
\name{read_data_package_report_acl}
\alias{read_data_package_report_acl}
\title{Read data package report access control list}
\usage{
read_data_package_report_acl(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) The access control list
}
\description{
Read data package report access control list
}
\note{
User authentication is required (see \code{login()})
}
\examples{


}
