% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_report_checksum.R
\name{read_data_package_report_checksum}
\alias{read_data_package_report_checksum}
\title{Read data package report checksum}
\usage{
read_data_package_report_checksum(package.id, environment = 
    'production')
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Data package report checksum
}
\description{
Read Data Package Report Checksum operation, specifying the scope, 
    identifier, and revision of the data package report object whose 
    checksum is to be read in the URI, returning a 40 character SHA-1 
    checksum value.
}
