% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_report_checksum.R
\name{read_data_package_report_checksum}
\alias{read_data_package_report_checksum}
\title{Read data package report checksum}
\usage{
read_data_package_report_checksum(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(character) A 40 character SHA-1 checksum value for the report
}
\description{
Read data package report checksum
}
\examples{
# Read report checksum
checksum <- read_data_package_report_checksum(packageId = "knb-lter-luq.208.1")
checksum

}
