% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_report_resource_metadata.R
\name{read_data_package_report_resource_metadata}
\alias{read_data_package_report_resource_metadata}
\title{Read data package report resource metadata}
\usage{
read_data_package_report_resource_metadata(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) Resource metadata
}
\description{
Read data package report resource metadata
}
\examples{
read_data_package_report_resource_metadata("knb-lter-mcm.9129.3")

}
