% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_report_resource_metadata.R
\name{read_data_package_report_resource_metadata}
\alias{read_data_package_report_resource_metadata}
\title{Read data package report resource metadata}
\usage{
read_data_package_report_resource_metadata(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Report resource metadata
}
\description{
Read data package report resource metadata
}
\examples{
# Read resource metadata
resourceMetadata <- read_data_package_report_resource_metadata(
  packageId = "knb-lter-mcm.9129.3")
resourceMetadata

}
