% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_report_summary.R
\name{read_data_package_report_summary}
\alias{read_data_package_report_summary}
\title{Summarize the data package quality report}
\usage{
read_data_package_report_summary(
  packageId,
  with_exceptions = TRUE,
  env = "production"
)
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{with_exceptions}{(logical) Convert quality report warnings and errors
to R warnings and errors}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(message/warning/error) A message listing the total number of checks
resulting in valid, info, warn, and error status. Exceptions are raised if
warnings and errors are found and \code{with_exceptions} is TRUE.
}
\description{
Summarize the data package quality report
}
\examples{
# Read report summary
read_data_package_report_summary("knb-lter-knz.260.4")
}
