% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_resource_metadata.R
\name{read_data_package_resource_metadata}
\alias{read_data_package_resource_metadata}
\title{Read data package resource metadata}
\usage{
read_data_package_resource_metadata(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(xml_document) Resource metadata of \code{packageId}
}
\description{
Read data package resource metadata
}
\examples{
\dontrun{

# Read resource metadata
resourceMetadata <- read_data_package_resource_metadata(
 packageId = "edi.613.1"
)
resourceMetadata
#> {xml_document}
#> <resourceMetadata>
#>  [1] <dateCreated>2021-06-08 18:02:43.361</dateCreated>
#>  [2] <doi>doi:10.6073/pasta/47d7cb6d374b6662cce98e42122169f8</doi>
#>  [3] <entityId>null</entityId>
#>  [4] <entityName>null</entityName>
#>  [5] <fileName>null</fileName>
#>  [6] <identifier>613</identifier>
#>  [7] <packageId>edi.613.1</packageId>
#>  [8] <principalOwner>uid=EDI,o=EDI,dc=edirepository,dc=org</principal ...
#>  [9] <resourceId>https://pasta.lternet.edu/package/eml/edi/613/1</res ...
#> [10] <resourceType>dataPackage</resourceType>
#> [11] <revision>1</revision>
#> [12] <scope>edi</scope>
}
}
\seealso{
Other Accessing: 
\code{\link{read_data_entity_checksum}()},
\code{\link{read_data_entity_names}()},
\code{\link{read_data_entity_name}()},
\code{\link{read_data_entity_resource_metadata}()},
\code{\link{read_data_entity_sizes}()},
\code{\link{read_data_entity_size}()},
\code{\link{read_data_entity}()},
\code{\link{read_data_package_archive}()},
\code{\link{read_data_package_doi}()},
\code{\link{read_data_package_error}()},
\code{\link{read_data_package_from_doi}()},
\code{\link{read_data_package_report_checksum}()},
\code{\link{read_data_package_report_resource_metadata}()},
\code{\link{read_data_package_report_summary}()},
\code{\link{read_data_package_report}()},
\code{\link{read_data_package}()},
\code{\link{read_evaluate_report_summary}()},
\code{\link{read_evaluate_report}()},
\code{\link{read_metadata_checksum}()},
\code{\link{read_metadata_dublin_core}()},
\code{\link{read_metadata_entity}()},
\code{\link{read_metadata_format}()},
\code{\link{read_metadata_resource_metadata}()},
\code{\link{read_metadata}()}
}
\concept{Accessing}
