% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data_package_resource_metadata.R
\name{read_data_package_resource_metadata}
\alias{read_data_package_resource_metadata}
\title{Read data package resource metadata}
\usage{
read_data_package_resource_metadata(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Resource metadata of \code{packageId}
}
\description{
Read data package resource metadata
}
\examples{


}
