% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_evaluate_report.R
\name{read_evaluate_report}
\alias{read_evaluate_report}
\title{Read evaluate report}
\usage{
read_evaluate_report(transaction, format = "xml", env = "production")
}
\arguments{
\item{transaction}{(character) Transaction identifier}

\item{format}{(character) Format of the returned report. Can be: "xml", "html", or "character".}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_document or html_document or character) The evaluate quality report document
}
\description{
Read evaluate report
}
\note{
User authentication is required (see \code{login()})
}
\examples{
# Evaluate data package
path <- "/Users/me/Documents/edi.468.1.xml"
transaction <- evaluate_data_package(path)

# Result in XML format
qualityReport <- read_evaluate_report(transaction)

# Result in HTML format
qualityReport <- read_evaluate_report(transaction, format = "html")

# Result as character string
qualityReport <- read_evaluate_report(transaction, format = "character")
message(qualityReport)
writeLines(qualityReport, "/Users/me/Documents/report.txt"))

}
