% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_evaluate_report.R
\name{read_evaluate_report}
\alias{read_evaluate_report}
\title{Read evaluate report}
\usage{
read_evaluate_report(transaction, html = FALSE, tier = "production")
}
\arguments{
\item{transaction}{(character) Transaction identifier returned for each data package evaluate, upload, and delete operation}

\item{html}{(logical) Return result in HTML format?}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) The evaluate quality report document
}
\description{
Read evaluate report
}
\examples{

}
