% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_evaluate_report.R
\name{read_evaluate_report}
\alias{read_evaluate_report}
\title{Read evaluate report}
\usage{
read_evaluate_report(package.id, identifier, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{identifier}{(character) Data entity identifier (e.g. 
5c224a0e74547b14006272064dc869b1)}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Data entity checksum
}
\description{
Read Evaluate Report operation, specifying the transaction identifier of the evaluate quality report document to be read in the URI. 

If an HTTP Accept header with value ‘text/html’ is included in the request, returns an HTML representation of the report. The default representation is XML. See the Evaluate Data Package service method for information about how to obtain the transaction identifier.
}
\details{
GET : https://pasta.lternet.edu/package/evaluate/report/eml/{transaction}
}
\examples{

}
