% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_evaluate_report.R
\name{read_evaluate_report}
\alias{read_evaluate_report}
\title{Read evaluate report}
\usage{
read_evaluate_report(transaction, html = FALSE, tier = "production")
}
\arguments{
\item{transaction}{(character) Transaction identifier}

\item{html}{(logical) Return result in HTML format?}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) The evaluate quality report document
}
\description{
Read evaluate report
}
\note{
User authentication is required (see \code{login()})
}
\examples{
# Evaluate data package
path <- "/Users/me/Documents/edi.468.1.xml"
transaction <- evaluate_data_package(path)

# Result in XML format
read_evaluate_report(transaction)

# Result in HTML format
read_evaluate_report(transaction, html = TRUE)

}
