% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata}
\usage{
read_metadata(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(xml_document) EML document
}
\description{
Read metadata
}
\examples{
read_metadata("edi.100.1")

}
