% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata}
\usage{
read_metadata(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(xml_document) EML metadata document
}
\description{
Read metadata
}
\examples{
# Read metadata
eml <- read_metadata("edi.100.1")
eml
}
