% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Read metadata}
\usage{
read_metadata(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('xml_document' 'xml_node') EML metadata.
}
\description{
Description Read Metadata (EML) operation, specifying the scope, identifier, and revision of the EML document to be read in the URI. Revision may be specified as “newest” or “oldest” to retrieve the newest or oldest revision, respectively.
}
\details{
GET : https://pasta.lternet.edu/package/metadata/eml/{scope}/{identifier}/{revision}
}
\examples{

}
