% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata_acl.R
\name{read_metadata_acl}
\alias{read_metadata_acl}
\title{Read metadata ACL}
\usage{
read_metadata_acl(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Data package report checksum
}
\description{
Read Metadata ACL operation, specifying the scope, identifier, and revision of the data package metadata whose Access Control List (ACL) is to be read in the URI, returning an XML string representing the ACL for the data package metadata resource. Please note: only a very limited set of users are authorized to use this service method.
}
\details{
GET : https://pasta.lternet.edu/package/metadata/acl/eml/{scope}/{identifier}/{revision}
}
\examples{

}
