% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata_acl.R
\name{read_metadata_acl}
\alias{read_metadata_acl}
\title{Read metadata Access Control List}
\usage{
read_metadata_acl(packageId, tier = "production", strip_ns = TRUE)
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}

\item{strip_ns}{(logical) Strip result namespace?}
}
\value{
(xml_document) Metadata ACL resource. Please note: only a very limited set of users are authorized to use this service method.
}
\description{
Read metadata Access Control List
}
\examples{

}
