% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata_checksum.R
\name{read_metadata_checksum}
\alias{read_metadata_checksum}
\title{Read metadata checksum}
\usage{
read_metadata_checksum(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
(character) Data package report checksum
}
\description{
Read Metadata Checksum operation, specifying the scope, identifier, and revision of the metadata object whose checksum value is to be read in the URI, returning a 40 character SHA-1 checksum value.
}
\details{
GET : https://pasta.lternet.edu/package/metadata/checksum/eml/{scope}/{identifier}/{revision}
}
\examples{

}
