% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata_checksum.R
\name{read_metadata_checksum}
\alias{read_metadata_checksum}
\title{Read metadata checksum}
\usage{
read_metadata_checksum(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(character) A 40 character SHA-1 checksum value
}
\description{
Read metadata checksum
}
\examples{
read_metadata_checksum("knb-lter-ntl.409.1")

}
