% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata_checksum.R
\name{read_metadata_checksum}
\alias{read_metadata_checksum}
\title{Read metadata checksum}
\usage{
read_metadata_checksum(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(character) A 40 character SHA-1 checksum value
}
\description{
Read metadata checksum
}
\examples{
# Read checksum
checksum <- read_metadata_checksum("knb-lter-ntl.409.1")
checksum

}
