% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata_dublin_core.R
\name{read_metadata_dublin_core}
\alias{read_metadata_dublin_core}
\title{Read metadata Dublin Core}
\usage{
read_metadata_dublin_core(package.id, environment = "production")
}
\arguments{
\item{package.id}{(character) Package identifier composed of scope, identifier, and
revision (e.g. 'edi.101.1').}

\item{environment}{(character) Data repository environment to create the package in.
Can be: 'development', 'staging', 'production'.}
}
\value{
('xml_document' 'xml_node') Dublin Core metadata
}
\description{
Read Metadata (Dublin Core) operation, specifying the scope, identifier, and revision of the Dublin Core metadata to be read in the URI. Revision may be specified as “newest” or “oldest” to retrieve the newest or oldest revision, respectively.
}
\details{
GET : https://pasta.lternet.edu/package/metadata/dc/{scope}/{identifier}/{revision}
}
\examples{

}
