% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata_dublin_core.R
\name{read_metadata_dublin_core}
\alias{read_metadata_dublin_core}
\title{Read metadata Dublin Core}
\usage{
read_metadata_dublin_core(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Dublin Core metadata
}
\description{
Read metadata Dublin Core
}
\examples{
# Read dc metadata
dc <- read_metadata_dublin_core("knb-lter-nes.10.1")
dc

}
