% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata_entity.R
\name{read_metadata_entity}
\alias{read_metadata_entity}
\title{Read data entity metadata}
\usage{
read_metadata_entity(packageId, entityId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{entityId}{(character) Data entity identifier}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
(xml_nodeset) The metadata of \code{entityId} in \code{packageId}
}
\description{
Read data entity metadata
}
\examples{
# Read entity names and IDs
packageId <- "knb-lter-cap.691.2"
entities <- read_data_entity_names(packageId)
entities

# Read metadata of the first entity
meta <- read_metadata_entity(packageId, entityId = entities$entityId[1])
meta

}
