% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata_format.R
\name{read_metadata_format}
\alias{read_metadata_format}
\title{Read metadata format}
\usage{
read_metadata_format(packageId, env = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(character) Metadata format type
}
\description{
Read metadata format
}
\examples{
# Read format
metadataFormat <- read_metadata_format("knb-lter-nwt.930.1")
metadataFormat
}
