% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata_format.R
\name{read_metadata_format}
\alias{read_metadata_format}
\title{Read metadata format}
\usage{
read_metadata_format(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier of the form "scope.identifier.revision"}

\item{tier}{(character) Repository tier, which can be: "production", "staging", or "development"}
}
\value{
(character) metadata format type
}
\description{
Read metadata format
}
\examples{
read_metadata_format("knb-lter-nwt.930.1")

}
