% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata_resource_metadata.R
\name{read_metadata_resource_metadata}
\alias{read_metadata_resource_metadata}
\title{Read metadata resource metadata}
\usage{
read_metadata_resource_metadata(packageId, tier = "production")
}
\arguments{
\item{packageId}{(character) Data package identifier}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Resource metadata for the data package metadata resource
}
\description{
Read metadata resource metadata
}
\examples{
read_metadata_resource_metadata("knb-lter-pal.309.1")

}
