% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tables.R
\name{read_tables}
\alias{read_tables}
\title{Read data tables of a data package}
\usage{
read_tables(eml)
}
\arguments{
\item{eml}{(xml_document, xml_node) EML metadata returned from \code{read_eml()}}
}
\value{
(list) List of named data frames
}
\description{
Read data tables of a data package from an EML metadata based repository. The EML contains all information required to do this.
}
\examples{

}
