% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_data_packages.R
\name{search_data_packages}
\alias{search_data_packages}
\title{Search data packages}
\usage{
search_data_packages(query, env = "production")
}
\arguments{
\item{query}{(character) Query (see details below)}

\item{env}{(character) Repository environment. Can be: "production",
"staging", or "development".}
}
\value{
(xml_document) Search results containing the fields:
\itemize{
  \item abstract
  \item begindate
  \item doi
  \item enddate
  \item funding
  \item geographicdescription
  \item id
  \item methods
  \item packageid
  \item pubdate
  \item responsibleParties
  \item scope
  \item site
  \item taxonomic
  \item title
  \item authors
  \item spatialCoverage
  \item sources
  \item keywords
  \item organizations
  \item singledates
  \item timescales
}
}
\description{
Searches data packages in the EDI data repository using the
specified Solr query.
}
\details{
Documents in the EDI data repository Solr index can be discovered
based on metadata values stored in the following list of searchable fields
(not all EML content is queryable):

Single-value fields:
\itemize{
  \item abstract
  \item begindate
  \item doi
  \item enddate
  \item funding
  \item geographicdescription
  \item id
  \item methods
  \item packageid
  \item pubdate
  \item responsibleParties
  \item scope
  \item singledate
  \item site
  \item taxonomic
  \item title
}

Multi-value fields:
\itemize{
  \item author
  \item coordinates
  \item keyword
  \item organization
  \item projectTitle
  \item relatedProjectTitle
  \item timescale
}

\code{query} parser: The optimal query parser (defType=edismax) is added to
every query.

See \href{https://cwiki.apache.org/confluence/display/solr/}{Apache Solr Wiki} for how to construct a Solr query.
}
\note{
Only the newest version of data packages are searchable, older
versions are not.

When constructing a query note that the 15403 data packages of the
\href{https://lternet.edu/the-ecotrends-project/}{ecotrends} project and
10492 data packages of the
\href{https://lternet.edu/lter-remote-sensing-and-geographic-information-system-data/}{LTER Landsat}
project, can be excluded from the returned results by including
\code{&fq=-scope:(ecotrends+lter-landsat)} in the query string.
}
\examples{
\dontrun{

# Search for data packages containing the term "air temperature"
search_data_packages(query = 'q="air+temperature"&fl=*')
#> {xml_document}
#> <resultset numFound="1347" start="0" rows="10">
#>  [1] <document>\n  <abstract>This project contains data collected fro ...
#>  [2] <document>\n  <abstract>This data set contains measurements of a ...
#>  [3] <document>\n  <abstract>Air temperature is measured at 15-minute ...
#>  [4] <document>\n  <abstract>This data package contains 5-minute air  ...
#>  [5] <document>\n  <abstract>This data package contains 5-minute air  ...
#>  [6] <document>\n  <abstract>This data package contains 5-minute air  ...
#>  [7] <document>\n  <abstract>This data package contains 5-minute air  ...
#>  [8] <document>\n  <abstract>This data package contains 5-minute air  ...
#>  [9] <document>\n  <abstract>This data package contains 5-minute air  ...
#> [10] <document>\n  <abstract>This data package contains 5-minute air  ...

# Search for data packages containing the term "air temperature" and
# returning only the packageid, title, and score of each match
search_data_packages(query = 'q="air+temperature"&fl=packageid,title,score')
#> {xml_document}
#> <resultset numFound="1347" start="0" rows="10">
#>  [1] <document>\n  <packageid>knb-lter-cwt.4038.19</packageid>\n  <ti ...
#>  [2] <document>\n  <packageid>knb-lter-cwt.4100.19</packageid>\n  <ti ...
#>  [3] <document>\n  <packageid>knb-lter-hbr.322.1</packageid>\n  <titl ...
#>  [4] <document>\n  <packageid>knb-lter-jrn.210548010.21</packageid>\n ...
#>  [5] <document>\n  <packageid>knb-lter-jrn.210548011.23</packageid>\n ...
#>  [6] <document>\n  <packageid>knb-lter-jrn.210548012.22</packageid>\n ...
#>  [7] <document>\n  <packageid>knb-lter-jrn.210548013.22</packageid>\n ...
#>  [8] <document>\n  <packageid>knb-lter-jrn.210548014.21</packageid>\n ...
#>  [9] <document>\n  <packageid>knb-lter-jrn.210548015.21</packageid>\n ...
#> [10] <document>\n  <packageid>knb-lter-jrn.210548001.22</packageid>\n ...

# Search for data packages containing the term "air temperature", returning
# only the packageid, title, score, and excluding ecotrends and lter-landsat
# scopes from the returned results
query <- paste0('q="air+temperature"&fl=packageid,title,score&',
                'fq=-scope:(ecotrends+lter-landsat)')
search_data_packages(query)
#> {xml_document}
#> <resultset numFound="917" start="0" rows="10">
#>  [1] <document>\n  <packageid>knb-lter-cwt.4038.19</packageid>\n  <ti ...
#>  [2] <document>\n  <packageid>knb-lter-cwt.4100.19</packageid>\n  <ti ...
#>  [3] <document>\n  <packageid>knb-lter-hbr.322.1</packageid>\n  <titl ...
#>  [4] <document>\n  <packageid>knb-lter-jrn.210548010.21</packageid>\n ...
#>  [5] <document>\n  <packageid>knb-lter-jrn.210548011.23</packageid>\n ...
#>  [6] <document>\n  <packageid>knb-lter-jrn.210548012.22</packageid>\n ...
#>  [7] <document>\n  <packageid>knb-lter-jrn.210548013.22</packageid>\n ...
#>  [8] <document>\n  <packageid>knb-lter-jrn.210548014.21</packageid>\n ...
#>  [9] <document>\n  <packageid>knb-lter-jrn.210548015.21</packageid>\n ...
#> [10] <document>\n  <packageid>knb-lter-jrn.210548001.22</packageid>\n ...
}
}
\concept{Browse and Discovery}
