% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_data_packages.R
\name{search_data_packages}
\alias{search_data_packages}
\title{Search data packages}
\usage{
search_data_packages(query, tier = "production")
}
\arguments{
\item{query}{(character) Query (see details below)}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
(xml_document) Search results
}
\description{
Searches data packages in the EDI Data Repository using the specified Solr query. See ADD RESOURCE for constructing queries.
}
\details{
Documents in the EDI Data Repository Solr index can be discovered based on metadata values stored in the following list of searchable fields (not all EML content is queryable):

Single-value fields:
\itemize{
  \item abstract
  \item begindate
  \item doi
  \item enddate
  \item funding
  \item geographicdescription
  \item id
  \item methods
  \item packageid
  \item pubdate
  \item responsibleParties
  \item scope
  \item singledate
  \item site
  \item taxonomic
  \item title
}

Multi-value fields:
\itemize{
  \item author
  \item coordinates
  \item keyword
  \item organization
  \item projectTitle
  \item relatedProjectTitle
  \item timescale
}

\code{query} parser: The optimal query parser (defType=edismax) is added to every query.

 See \href{https://cwiki.apache.org/confluence/display/solr/}{Apache Solr Wiki} for how to construct a Solr query.
}
\note{
Only the newest version of data packages are searchable, older versions are not.

When constructing a query note that the 15403 data packages of the \href{https://lternet.edu/the-ecotrends-project/}{ecotrends} project and 10492 data packages of the \href{https://lternet.edu/lter-remote-sensing-and-geographic-information-system-data/}{LTER Landsat} project, can be excluded from the returned results by including \code{&fq=-scope:(ecotrends+lter-landsat)} in the query string.
}
\examples{
# Search for data packages containing the term "air temperature"
query <- 'q="air+temperature"&fl=*'
search_data_packages(query)

# Search for data packages containing the term "air temperature" and returning 
# only the packageid, title, and score of each match
query <- 'q="air+temperature"&fl=packageid,title,score'
search_data_packages(query)

# Search for data packages containing the term "air temperature", returning 
# only the packageid, title, score, and excluding ecotrends and lter-landsat 
# scopes from the returned results
query <- 'q="air+temperature"&fl=packageid,title,score&fq=-scope:(ecotrends+lter-landsat)'
search_data_packages(query)

}
