% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_package.R
\name{update_data_package}
\alias{update_data_package}
\title{Update data package}
\usage{
update_data_package(eml, useChecksum = FALSE, env = "production")
}
\arguments{
\item{eml}{(character) Full path to an EML file describing the data package to be updated}

\item{useChecksum}{(logical) Use data entities from a previous version of the data package? See details below.}

\item{env}{(character) Repository environment. Can be: "production", "staging", or "development".}
}
\value{
transaction (character) Transaction identifier. May be used in a subsequent call to:
\itemize{
  \item \code{check_status_update()} to determine the operation status
  \item \code{read_data_package()} to obtain the data package resource map
}
}
\description{
Update data package
}
\details{
Each data entity described in \code{eml} must be accompanied by a web accessible URL at the XPath ".//physical/distribution/online/url". The EDI data repository uses these links to download the data entities. The URLs must be static and not have any redirects otherwise the data entities will not be downloadable.
}
\note{
User authentication is required (see \code{login()})
}
\examples{
path <- "/Users/me/Documents/edi.468.2.xml"
update_data_package(path)


}
