% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data_package.R
\name{update_data_package}
\alias{update_data_package}
\title{Update data package}
\usage{
update_data_package(eml, useChecksum = FALSE, tier = "production")
}
\arguments{
\item{eml}{(character) Full path to an EML file}

\item{useChecksum}{(logical) Use data entities from a previous version of the data package? See details below.}

\item{tier}{(character) Repository tier. Can be: "production", "staging", or "development".}
}
\value{
transaction (character) Transaction identifier. Use this value with: 
\itemize{
  \item \code{check_status_update()} to see if the update has completed or if any errors occurred while processing the request
  \item \code{read_evaluate_report()} to read the evaluation report
  \item \code{read_data_package()} to obtain the data package resource map
}
}
\description{
Update data package
}
\details{
Each data entity described in \code{eml} must be accessible to the EDI Data Repository through URLs listed in the \code{eml} at the XPath ".//physical/distribution/online/url". This URL should be static and not present the EDI Data Repository with any redirects, otherwise the data entity will not be accessible.
}
\examples{
path <- "/Users/me/Documents/edi.468.2.xml"
update_data_package(path)


}
