% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_arguments.R
\name{validate_arguments}
\alias{validate_arguments}
\title{Validate arguments to EDIutils functions}
\usage{
validate_arguments(x)
}
\arguments{
\item{x}{(list) A named list of arguments and corresponding values.}
}
\description{
Validate arguments to EDIutils functions
}
\details{
Checks performed by this function provide user friendly warnings and error messages to make their lives simpler.
}
